% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdatamoments.R
\docType{class}
\name{cdatamoments-class}
\alias{cdatamoments-class}
\alias{.cdatamoments}
\title{Finmix \code{cdatamoments} class}
\description{
Stores moments of an \linkS4class{fdata} object containing continuous data.
The \code{fdata} object is stored in the parent \linkS4class{datamoments}
class.
}
\section{Slots}{

\describe{
\item{\code{higher}}{An array containing the four higher centralized moments of the
continuous data stored in the \code{fdata} object.}

\item{\code{skewness}}{A vector storing the skewness of the continuous data in the
corresponding \code{fdata} object.}

\item{\code{kurtosis}}{A vector storing the kurtosis of the continuous data in the
corresponding \code{fdata} object.}

\item{\code{corr}}{A matrix containing the correlations between the data dimensions
in case of multivariate data (i.e. slot \code{r} in the \code{fdata} object is
larger than one).}

\item{\code{smoments}}{A \code{csdatamoments} object, if the \code{fdata} object also holds
indicators. \code{NULL}, if no indicators are present in the \code{fdata} object.}
}}

\seealso{
\itemize{
\item \linkS4class{datamoments} for the parent class
\item \linkS4class{ddatamoments} for the corresponding class for
discrete data
\item \linkS4class{csdatamoments} for the contained class if indicators
are present in the \code{fdata} object
}
}
\keyword{internal}
