% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmodelmoments.R
\docType{class}
\name{cmodelmoments}
\alias{cmodelmoments}
\alias{.cmodelmoments}
\title{Finmix \code{cmodelmoments} class}
\description{
This class defines the general theoretical moments of a finite mixture model
with continuous data.
}
\section{Slots}{

\describe{
\item{\code{higher}}{An array containing the four higher centralized moments of the
(in case of multivariate data marginal) finite mixture.}

\item{\code{skewness}}{A vector containing the skewness(es) of the finite mixture
model.}

\item{\code{kurtosis}}{A vector containing the kurtosis(es) of the finite mixture
model.}
}}

\seealso{
\itemize{
\item \link{modelmoments} for the base class
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of any \code{modelmoments} inherited class
}
}
