% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdatamoments.R
\docType{class}
\name{csdatamoments-class}
\alias{csdatamoments-class}
\alias{.csdatamoments}
\title{Finmix \code{csdatamoments} class}
\description{
Stores moments for indicators of continuous data. Inherited directly from
the \linkS4class{sdatamoments} class.
}
\section{Slots}{

\describe{
\item{\code{B}}{A vector storing the between-group heterogeneity.}

\item{\code{W}}{A vector storing the within-group heterogeneity.}

\item{\code{T}}{A vector storing the total variance.}

\item{\code{R}}{A numeric storing the coefficient of determination for univariate
data.}

\item{\code{Rdet}}{A numeric storing the coefficient of determination using the
trace for multivariate data.}

\item{\code{Rtr}}{A numeric storing the coefficient of determination using the
determinants for multivariate data.}
}}

\seealso{
\itemize{
\item \linkS4class{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
\item \linkS4class{csdatamoments} for the corresponding class defining
moments for data from a discrete-valued finite mixture
}
}
\keyword{internal}
