% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdatamoments.R
\docType{class}
\name{csdatamomentsOrNULL-class}
\alias{csdatamomentsOrNULL-class}
\title{Finmix class union of \code{csdatamoments} and \code{NULL}}
\description{
Defines a class union such that the object held by a child class can also
be \code{NULL}.
}
\keyword{internal}
