% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclass.R
\docType{class}
\name{dataclass_class}
\alias{dataclass_class}
\alias{.dataclass}
\title{Finmix \code{dataclass} class}
\description{
Stores objects to classify observations using a fully specified mixture
model. If the indicators a finite mixture model is fully specified as then
the likelihood can be calculated for each observation depending on the
component it stems from.
}
\section{Slots}{

\describe{
\item{\code{logpy}}{An array containing the logarithmized}

\item{\code{prob}}{An array storing the probability classification matrix that
defines for each observation the probability of belonging to component
\code{k}. Henceforth, each row sums to one. The dimension of this array is
\verb{N x K}.}

\item{\code{mixlik}}{A numeric storing the logarithm of the mixture likelihood
evaluated at certain parameters \code{par} from a finmix \code{model} object and
corresponding \code{weights}.}

\item{\code{entropy}}{A numeric storing the entropy of the classification.}

\item{\code{loglikcd}}{An array storing the logarithm of the conditional likelihood
of each component parameter, if indicators have not been simulated. The
array has dimension \verb{1 x K}.}

\item{\code{postS}}{A numeric storing the posterior probability of the indicators
\code{S} in the data, if indicators have been simulated.}
}}

\references{
Frühwirth-Schnatter, S. (2006), "Finite Mixture and Markov Switching Models"
}
\seealso{
\itemize{
\item \link[=fdata_class]{fdata} for the class holding the data
\item \link[=model_class]{model} for the class defining a finite mixture model
\item \code{\link[=dataclass]{dataclass()}} for the constructor of this class
}
}
