% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamoments.R
\docType{class}
\name{datamoments-class}
\alias{datamoments-class}
\alias{.datamoments}
\title{Finmix \code{datamoments} class}
\description{
Stores moments of a corresponding \code{fdata} object.
}
\section{Slots}{

\describe{
\item{\code{mean}}{A numeric storing the mean of the slot \code{y} in the \code{fdata} object.}

\item{\code{var}}{A matrix storing the variance(s and covariances) of the \code{y} slot
in the \code{fdata} object.}

\item{\code{fdata}}{An \code{fdata} object containing the observations and possible
indicators.}
}}

\seealso{
\itemize{
\item \linkS4class{cdatamoments} for data moments of continuous data
\item \linkS4class{ddatamoments} for data moments of discrete data
\item \linkS4class{sdatamoments} for data moments of the indicators
}
}
