% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamoments.R
\name{datamoments}
\alias{datamoments}
\title{Constructor for \code{datamoments} classes}
\usage{
datamoments(value = fdata())
}
\arguments{
\item{value}{An \code{fdata} object with at least slot \code{y} non-empty.}
}
\value{
A \code{datamoments} object containing the data moments for the
observations in slot \code{y} and if available also for the indicators in slot
\code{S}.
}
\description{
Calling \code{\link[=datamoments]{datamoments()}} generates the moments for an \code{fdata} object.
Depending on the type of data either an \code{cdatamoments} or \code{ddatamoments}
object is generated. If in addition the \code{fdata} object contains fixed
indicators, the \code{datamoments} object also holds an \code{sdatamoments} object to
store the data moments of these indicators.
}
\examples{
# Create an fdata class with Poisson data.
f_data <- fdata(rpois(100, 312), sim=TRUE)
# Compute the data moments.
datamoments(f_data)

}
\seealso{
\itemize{
\item \linkS4class{datamoments} for all slots of this class
\item \linkS4class{cdatamoments} for the class for continuous data
\item \linkS4class{ddatamoments} for the class for discrete data
}
}
