% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normultmodelmoments.R, R/csdatamoments.R,
%   R/datamoments.R
\docType{class}
\name{getB,normultmodelmoments-method}
\alias{getB,normultmodelmoments-method}
\alias{getB,csdatamoments-method}
\alias{datamoments_class}
\alias{.datamoments}
\title{Getter method of \code{normultmodelmoments} class.}
\usage{
\S4method{getB}{normultmodelmoments}(object)

\S4method{getB}{csdatamoments}(object)
}
\arguments{
\item{object}{An \code{csdatamoments} object.}
}
\value{
The \code{B} slot of the \code{object}.

The \code{B} slot of the \code{object}.
}
\description{
Returns the \code{B} slot.

Returns the \code{B} slot.

Stores moments of a corresponding \code{fdata} object.
}
\section{Functions}{
\itemize{
\item \code{getB,normultmodelmoments-method}: Getter method for slot \code{B}

\item \code{getB,csdatamoments-method}: 
}}

\section{Slots}{

\describe{
\item{\code{mean}}{A numeric storing the mean of the slot \code{y} in the \code{fdata} object.}

\item{\code{var}}{A matrix storing the variance(s and covariances) of the \code{y} slot
in the \code{fdata} object.}

\item{\code{VIRTUAL}}{Virtual class containing further data moments.}
}}

\examples{
f_model         <- model("normult", weight = matrix(c(.3, .7), nrow = 1))
means           <- matrix(c(-2, -2, 2, 2),nrow = 2)
covar           <- matrix(c(1, 1.2, 1.2, 4), nrow = 2)
sigmas          <- array(c(covar, 2*covar), dim = c(2, 2, 2))
setPar(f_model) <- list(mu = means, sigma = sigmas)
f_moments       <- modelmoments(f_model)
getB(f_moments)

# Generate an exponential mixture model with two components.
f_model <- model("exponential", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Calculate the mixture moments.
f_sdatamoms <- sdatamoments(f_data)
# Get the moments for the included indicators of the data. 
getB(f_sdatamoms)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}

\itemize{
\item \link[=datamoments_class]{datamoments} for the base class for model moments
\item \link[=datamoments]{datamoments()} for the constructor of the \code{datamoments}
class family
\item \link[=sdatamoments_class]{csdatamoments} for the class definition
\item \link[=sdatamoments]{sdatamoments()} for the constructor of the class
}

\itemize{
\item \link{cdatamoments} for data moments of continuous data
\item \link{ddatamoments} for data moments of discrete data
\item \link{sdatamoments} for data moments of the indicators
}
}
