% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddatamoments.R
\docType{class}
\name{ddatamoments-class}
\alias{ddatamoments-class}
\alias{.ddatamoments}
\title{Finmix \code{ddatamoments} class}
\description{
Stores moments of an \linkS4class{fdata} object containing discrete data.
The \code{fdata} object is stored in the parent \linkS4class{datamoments}
class.
}
\section{Slots}{

\describe{
\item{\code{factorial}}{An array containing the first four factorial moments of the
discrete data stored in the \code{fdata} object.}

\item{\code{over}}{A vector storing the overdispersion of the discrete data in the
corresponding \code{fdata} object.}

\item{\code{zero}}{A vector storing the fractions of zeros in the observed data. <}

\item{\code{smoments}}{An \code{sdatamoments} object, if the \code{fdata} object also holds
indicators. \code{NULL}, if no indicators are present in the \code{fdata} object.}
}}

\seealso{
\itemize{
\item \linkS4class{datamoments} for the parent class
\item \linkS4class{ddatamoments} for the corresponding class for
continuous data
\item \linkS4class{sdatamoments} for the contained class if indicators
are present in the \code{fdata} object
}
}
\keyword{internal}
