% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmodelmoments.R
\docType{class}
\name{dmodelmoments}
\alias{dmodelmoments}
\alias{.dmodelmoments}
\title{Finmix \code{dmodelmoments} class}
\description{
This class defines the general theoretical moments of a finite mixture model
with discrete data.
}
\section{Slots}{

\describe{
\item{\code{over}}{A numeric containing the over-dispersion.}

\item{\code{factorial}}{An array containing the first four factorial moments.}

\item{\code{zero}}{An numeric cotaining the excess zeros.}
}}

\seealso{
\itemize{
\item \link{modelmoments} for the base class
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of any \code{modelmoments} inherited class
}
}
