% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmodelmoments.R
\name{.generateMomentsNormal}
\alias{.generateMomentsNormal}
\title{Generate model moments for an normal mixture}
\usage{
.generateMomentsNormal(object)
}
\arguments{
\item{object}{An \code{normalmodelmoments} object to contain all calculated
moments.}
}
\value{
An \code{normalmodelmoments} object containing all moments of the
normal mixture distributions.
}
\description{
Only called implicitly. generates all moments of an normal mixture
distribution.
}
\keyword{internal}
