% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dstud}
\alias{dstud}
\title{Density function of a Student-t distribution}
\usage{
dstud(x, mu, sigma, df)
}
\arguments{
\item{x}{A vector of valued for which the density should be calculated.}

\item{mu}{A vector containing the mean of the distribution.}

\item{sigma}{A vector containing the standard deviation of the distribution.}

\item{df}{A vector containing the degrees of freedom of the distribution.}
}
\value{
The density of the Student-t distribution for the values of \code{x}.
}
\description{
Unused at this moment.
}
\keyword{internal}
