% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfix.R
\name{extract,mcmcoutputfix,numeric-method}
\alias{extract,mcmcoutputfix,numeric-method}
\title{Extracts samples from \code{mcmcoutput} object of a multivariate Normal mixture}
\usage{
\S4method{extract}{mcmcoutputfix,numeric}(object, index)
}
\arguments{
\item{object}{An \code{mcmcoutput} object from MCMC sampling of a multivariate
Normal mixture model.}

\item{index}{An numeric indicating which dimension of the multivariate
mixture should be extracted.}
}
\value{
An object class \code{mcmcextract} containing all samples of an extracted
dimension.
}
\description{
This function extracts samples from a multivariate Normal mixture output.
}
\keyword{internal}
