% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{extract}
\alias{extract}
\title{Extracts the MCMC samples from a specific dimension of a multivariate model}
\usage{
extract(object, index)
}
\description{
Extracts the MCMC samples from a specific dimension of a multivariate model
}
\keyword{internal}
