% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialmodelmoments.R
\name{generateMoments,binomialmodelmoments-method}
\alias{generateMoments,binomialmodelmoments-method}
\title{Generate moments for binomial mixture}
\usage{
\S4method{generateMoments}{binomialmodelmoments}(object)
}
\arguments{
\item{object}{An \code{binomialmodelmoments} object.}
}
\value{
An \code{binomialmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
binomial mixture distribution.
}
\keyword{internal}
