% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdatamoments.R
\name{generateMoments,cdatamoments-method}
\alias{generateMoments,cdatamoments-method}
\title{Generate moments for continuous data.}
\usage{
\S4method{generateMoments}{cdatamoments}(object)
}
\arguments{
\item{object}{An \code{cdatamoments} object.}
}
\value{
An \code{cdatamoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of a
finite mixture with continuous data.
}
\keyword{internal}
