% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdatamoments.R
\name{generateMoments,csdatamoments-method}
\alias{generateMoments,csdatamoments-method}
\title{Generate moments for indicators from a mixture with continuous data}
\usage{
\S4method{generateMoments}{csdatamoments}(object)
}
\arguments{
\item{object}{An \code{csdatamoments} object.}
}
\value{
An \code{csdatamoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of a
finite mixture with continuous data.
}
\keyword{internal}
