% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialmodelmoments.R
\name{generateMoments,exponentialmodelmoments-method}
\alias{generateMoments,exponentialmodelmoments-method}
\title{Generate moments for exponential mixture}
\usage{
\S4method{generateMoments}{exponentialmodelmoments}(object)
}
\arguments{
\item{object}{An \code{exponentialmodelmoments} object.}
}
\value{
An \code{exponentialmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
exponential mixture distribution.
}
\keyword{internal}
