% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmoments.R
\name{generateMoments,groupmoments-method}
\alias{generateMoments,groupmoments-method}
\title{Generate moments}
\usage{
\S4method{generateMoments}{groupmoments}(object)
}
\arguments{
\item{object}{A \code{groupmoments} object.}
}
\value{
An \code{groupmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of a
finite mixture with continuous data.
}
\keyword{internal}
