% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmodelmoments.R
\name{generateMoments,normalmodelmoments-method}
\alias{generateMoments,normalmodelmoments-method}
\title{Generate moments for normal mixture}
\usage{
\S4method{generateMoments}{normalmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normalmodelmoments} object.}
}
\value{
An \code{normalmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{generateMoments()} generates the moments of an
normal mixture distribution.
}
\keyword{internal}
