% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normultmodelmoments.R
\name{generateMoments,normultmodelmoments-method}
\alias{generateMoments,normultmodelmoments-method}
\title{Generate moments for normult mixture}
\usage{
\S4method{generateMoments}{normultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normultmodelmoments} object.}
}
\value{
An \code{normultmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
normult mixture distribution.
}
\keyword{internal}
