% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonmodelmoments.R
\name{generateMoments,poissonmodelmoments-method}
\alias{generateMoments,poissonmodelmoments-method}
\title{Generate moments for poisson mixture}
\usage{
\S4method{generateMoments}{poissonmodelmoments}(object)
}
\arguments{
\item{object}{An \code{poissonmodelmoments} object.}
}
\value{
An \code{poissonmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
poisson mixture distribution.
}
\keyword{internal}
