% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentmodelmoments.R
\name{generateMoments,studentmodelmoments-method}
\alias{generateMoments,studentmodelmoments-method}
\title{Generate moments for student mixture}
\usage{
\S4method{generateMoments}{studentmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studentmodelmoments} object.}
}
\value{
An \code{studentmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
student mixture distribution.
}
\keyword{internal}
