% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studmultmodelmoments.R
\name{generateMoments,studmultmodelmoments-method}
\alias{generateMoments,studmultmodelmoments-method}
\title{Generate moments for studmult mixture}
\usage{
\S4method{generateMoments}{studmultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studmultmodelmoments} object.}
}
\value{
An \code{studmultmodelmoments} object with calculated moments.
}
\description{
Implicit method. Calling \code{\link[=generateMoments]{generateMoments()}} generates the moments of an
studmult mixture distribution.
}
\keyword{internal}
