% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{generateMoments}
\alias{generateMoments}
\title{Generates the moments of a finite mixture model}
\usage{
generateMoments(object)
}
\description{
Generates the moments of a finite mixture model
}
\keyword{internal}
