% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/prior.R
\docType{methods}
\name{generatePrior}
\alias{generatePrior}
\alias{generatePrior,prior-method}
\title{Generates the prior for a specific \code{model}}
\usage{
generatePrior(object, ...)

\S4method{generatePrior}{prior}(object, fdata, model, varargin, prior.wagner, s)
}
\arguments{
\item{object}{A \code{prior} object to store the prior parameters and weights.}

\item{fdata}{An \code{fdata} object holding the data. Observations in slot \verb{@y}
must be existent.}

\item{model}{A \code{model} object specifying the finite mixture model.}

\item{varargin}{\code{NULL} or a \code{prior} object. This enables the user to pass in
an already constructed prior object that gets then completed.}

\item{prior.wagner}{A logical indicating, if the prior from Wagner (2007)
should be used in case of an exponential mixture model.}

\item{s}{A numeric specifying the standard deviation \code{s} for the
Metropolis-Hastings proposal.}
}
\description{
Calling \code{generatePrior()} generates the \code{prior} object when \code{\link[=priordefine]{priordefine()}}
had been called. When this function is called all checks have been passed
and \code{prior} construction can take place.
}
\seealso{
\itemize{
\item \linkS4class{prior} for the class definition
\item \code{\link[=priordefine]{priordefine()}} for the advanced class constructor using this method
}
}
\keyword{internal}
