% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialmodelmoments.R
\name{getB,exponentialmodelmoments-method}
\alias{getB,exponentialmodelmoments-method}
\title{Getter method of \code{exponentialmodelmoments} class.}
\usage{
\S4method{getB}{exponentialmodelmoments}(object)
}
\arguments{
\item{object}{An \code{exponentialmodelmoments} object.}
}
\value{
The \code{B} slot of the \code{object}.
}
\description{
Returns the \code{B} slot.
}
\examples{
f_model <- model("exponential", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getB(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
