% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normultmodelmoments.R
\name{getB,normultmodelmoments-method}
\alias{getB,normultmodelmoments-method}
\title{Getter method of \code{normultmodelmoments} class.}
\usage{
\S4method{getB}{normultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normultmodelmoments} object.}
}
\value{
The \code{B} slot of the \code{object}.
}
\description{
Returns the \code{B} slot.
}
\examples{
f_model         <- model("normult", weight = matrix(c(.3, .7), nrow = 1))
means           <- matrix(c(-2, -2, 2, 2),nrow = 2)
covar           <- matrix(c(1, 1.2, 1.2, 4), nrow = 2)
sigmas          <- array(c(covar, 2*covar), dim = c(2, 2, 2))
setPar(f_model) <- list(mu = means, sigma = sigmas)
f_moments       <- modelmoments(f_model)
getB(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
