% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{getBycolumn,fdata-method}
\alias{getBycolumn,fdata-method}
\title{Getter method of \code{fdata} class.}
\usage{
\S4method{getBycolumn}{fdata}(object)
}
\arguments{
\item{object}{An \code{fdata} object.}
}
\value{
The \code{bycolumn} slot of the \code{object}.
}
\description{
Returns the \code{bycolumn} slot of an \code{fdata} object.
}
\examples{
# Create an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
getBycolumn(f_data)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
