% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputbase.R
\name{getClust,mcmcoutputbase-method}
\alias{getClust,mcmcoutputbase-method}
\title{Getter method of \code{mcmcoutput} class.}
\usage{
\S4method{getClust}{mcmcoutputbase}(object)
}
\arguments{
\item{object}{An \code{mcmcoutput} object.}
}
\value{
The \code{clust} slot of the \code{object}.
}
\description{
Returns the \code{clust} slot.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Complete object slots for consistency. 
(f_data ~ f_model ~ f_mcmc) \%=\% mcmcstart(f_data, f_model, f_mcmc)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
# Get the slot.
getClust(f_output)

}
\seealso{
\itemize{
\item \link[=mcmcoutputbase-class]{mcmcoutput} for the class definition
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
}
}
\keyword{internal}
