% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getColExp}
\alias{getColExp}
\title{Getter for the \code{exp} slot in column format}
\usage{
getColExp(object)
}
\description{
Getter for the \code{exp} slot in column format
}
\keyword{internal}
