% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getColS}
\alias{getColS}
\title{Getter for the \code{S} slot in column format}
\usage{
getColS(object)
}
\description{
Getter for the \code{S} slot in column format
}
\keyword{internal}
