% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{getColT,fdata-method}
\alias{getColT,fdata-method}
\title{Getter method of \code{fdata} class.}
\usage{
\S4method{getColT}{fdata}(object)
}
\arguments{
\item{object}{An \code{fdata} object.}
}
\value{
The \code{T} slot of the \code{object} as a column-ordered matrix.
}
\description{
Returns the \code{T} slot as a column-ordered matrix.
}
\examples{
# Create an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
getColT(f_data)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
\keyword{internal}
