% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getColT}
\alias{getColT}
\title{Getter for the \code{T} slot in column format}
\usage{
getColT(object)
}
\description{
Getter for the \code{T} slot in column format
}
\keyword{internal}
