% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getColY}
\alias{getColY}
\title{Getter for the \code{y} slot in column format}
\usage{
getColY(object)
}
\description{
Getter for the \code{y} slot in column format
}
\keyword{internal}
