% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestind.R
\name{getEavg,mcmcestind-method}
\alias{getEavg,mcmcestind-method}
\title{Getter method of \code{mcmcestind} class.}
\usage{
\S4method{getEavg}{mcmcestind}(object)
}
\arguments{
\item{object}{An \code{mcmcestind} object.}
}
\value{
The \code{eavg} slot of the \code{object}.
}
\description{
Returns the \code{eavg} slot.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_est <- mcmcestimate(f_output)
# Get the slot.
getEavg(f_est)

}
\seealso{
\itemize{
\item \linkS4class{mcmcestfix} for the parent class with fixed indicators
\item \code{\link[=mcmcestimate]{mcmcestimate()}} for calculating point estimates from MCMC samples
}
}
\keyword{internal}
