% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getEntropyperm}
\alias{getEntropyperm}
\title{Getter for the \code{entropyperm} slot}
\usage{
getEntropyperm(object)
}
\description{
Getter for the \code{entropyperm} slot
}
\keyword{internal}
