% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getExtrabinvar}
\alias{getExtrabinvar}
\title{Getter for the \code{extrabinvar} slot}
\usage{
getExtrabinvar(object)
}
\description{
Getter for the \code{extrabinvar} slot
}
\keyword{internal}
