% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\name{getGmoments,sdatamoments-method}
\alias{getGmoments,sdatamoments-method}
\title{Getter method of \code{sdatamoments} class.}
\usage{
\S4method{getGmoments}{sdatamoments}(object)
}
\arguments{
\item{object}{An \code{sdatamoments} object.}
}
\value{
The \code{gmoments} slot of the \code{object}.
}
\description{
Returns the \code{gmoments} slot.
}
\examples{
# Generate a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Calculate the mixture moments.
f_sdatamoms <- sdatamoments(f_data)
# Get the moments for the included indicators of the data. 
getGmoments(f_sdatamoms)

}
\seealso{
\itemize{
\item \linkS4class{datamoments} for the base class for model moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of the \code{datamoments}
class family
\item \linkS4class{sdatamoments} for the class definition
\item \code{\link[=sdatamoments]{sdatamoments()}} for the constructor of the class
}
}
\keyword{internal}
