% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmodelmoments.R
\name{getHigher,cmodelmoments-method}
\alias{getHigher,cmodelmoments-method}
\title{Getter method of \code{cmodelmoments} class.}
\usage{
\S4method{getHigher}{cmodelmoments}(object)
}
\arguments{
\item{object}{An \code{cmodelmoments} object.}
}
\value{
The \code{higher} slot of the \code{object}.
}
\description{
Returns the \code{higher} slot.
}
\examples{
f_model <- model("exponential", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getHigher(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
