% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhier.R
\name{getHyper,mcmcoutputhier-method}
\alias{getHyper,mcmcoutputhier-method}
\title{Getter method of \code{mcmcoutputhier} class.}
\usage{
\S4method{getHyper}{mcmcoutputhier}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputhier} object.}
}
\value{
The \code{hyper} slot of the \code{object}.
}
\description{
Returns the \code{hyper} slot.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
# Get the slot.
getHyper(f_output)

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputhier} for the class definition
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
}
}
\keyword{internal}
