% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermindhier.R
\name{getHyperperm,mcmcpermfixpost-method}
\alias{getHyperperm,mcmcpermfixpost-method}
\title{Getter method of \code{mcmcperminfhier} class.}
\usage{
\S4method{getHyperperm}{mcmcpermfixpost}(object)
}
\arguments{
\item{object}{An \code{mcmcpermindhier} object.}
}
\value{
The \code{hyperperm} slot of the \code{object}.
}
\description{
Returns the \code{hyperperm} slot.
}
\examples{
\dontrun{getHyperpem(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputpermhier} for the inheriting class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
