% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfix.R
\name{getMperm,mcmcpermfix-method}
\alias{getMperm,mcmcpermfix-method}
\alias{mcmcpermfix_class,}
\alias{mcmcpermfixhier_class,}
\alias{mcmcpermfixpost_class,}
\alias{mcmcpermfixhierpost}
\title{Getter method of \code{mcmcpermfix} class.}
\usage{
\S4method{getMperm}{mcmcpermfix}(object)
}
\arguments{
\item{object}{An \code{mcmcpermfix} object.}
}
\value{
The \code{Mperm} slot of the \code{object}.
}
\description{
Returns the \code{Mperm} slot.
}
\examples{
\dontrun{getMperm(mcmcperm)}

}
\seealso{
\itemize{
\item \code{\link{mcmcoutputpermfix_class}} for the inheriting class
\item \code{\link{mcmcpermute}} for function permuting MCMC samples
}
}
