% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{getN,fdata-method}
\alias{getN,fdata-method}
\title{Getter method of \code{fdata} class.}
\usage{
\S4method{getN}{fdata}(object)
}
\arguments{
\item{object}{An \code{fdata} object.}
}
\value{
The \code{N} slot of the \code{object}.
}
\description{
Returns the \code{N} slot of an \code{fdata} object.
}
\examples{
# Create an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
getN(f_data)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
\keyword{internal}
