% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermind.R
\name{getNKperm,mcmcpermind-method}
\alias{getNKperm,mcmcpermind-method}
\title{Getter method of \code{mcmcpermind} class.}
\usage{
\S4method{getNKperm}{mcmcpermind}(object)
}
\arguments{
\item{object}{An \code{mcmcpermind} object.}
}
\value{
The \code{NKperm} slot of the \code{object}.
}
\description{
Returns the \code{NKperm} slot.
}
\examples{
\dontrun{getNKperm(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputpermbase} for the inheriting class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
