% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddatamoments.R
\name{getOver,ddatamoments-method}
\alias{getOver,ddatamoments-method}
\title{Getter method of \code{ddatamoments} class.}
\usage{
\S4method{getOver}{ddatamoments}(object)
}
\arguments{
\item{object}{An \code{ddatamoments} object.}
}
\value{
The \code{over} slot of the \code{object}.
}
\description{
Returns the \code{over} slot.
}
\examples{
# Generate a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Calculate the mixture moments.
f_datamoms <- datamoments(f_data)
# Get the moments for the included indicators of the data. 
getOver(f_datamoms)

}
\seealso{
\itemize{
\item \linkS4class{datamoments} for the base class for model moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of the \code{datamoments} class family
}
}
\keyword{internal}
