% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmodelmoments.R
\name{getOver,dmodelmoments-method}
\alias{getOver,dmodelmoments-method}
\title{Getter method of \code{dmodelmoments} class.}
\usage{
\S4method{getOver}{dmodelmoments}(object)
}
\arguments{
\item{object}{An \code{dmodelmoments} object.}
}
\value{
The \code{higher} slot of the \code{object}.
}
\description{
Returns the \code{higher} slot.
}
\examples{
f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getOver(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
