% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{getPar,model-method}
\alias{getPar,model-method}
\title{Getter method of \code{model} class.}
\usage{
\S4method{getPar}{model}(object)
}
\arguments{
\item{object}{An \code{model} object.}
}
\value{
The \code{par} slot of the \code{object}.
}
\description{
Returns the \code{par} slot.
}
\examples{
# Generate an exponential mixture model with two components.
f_model <- model("exponential", par = list(lambda = c(0.3, 0.7)), K = 2)
# Get the slot
getPar(f_model)

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
