% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfix.R
\name{getParperm,mcmcpermfix-method}
\alias{getParperm,mcmcpermfix-method}
\title{Getter method of \code{mcmcpermfix} class.}
\usage{
\S4method{getParperm}{mcmcpermfix}(object)
}
\arguments{
\item{object}{An \code{mcmcpermfix} object.}
}
\value{
The \code{parperm} slot of the \code{object}.
}
\description{
Returns the \code{parperm} slot.
}
\examples{
\dontrun{getParperm(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutput} for the inheriting class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
