% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfixpost.R
\name{getPostperm,mcmcpermfixpost-method}
\alias{getPostperm,mcmcpermfixpost-method}
\title{Getter method of \code{mcmcpermfixpost} class.}
\usage{
\S4method{getPostperm}{mcmcpermfixpost}(object)
}
\arguments{
\item{object}{An \code{mcmcpermfixpost} object.}
}
\value{
The \code{postperm} slot of the \code{object}.
}
\description{
Returns the \code{postperm} slot.
}
\examples{
\dontrun{getMperm(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputpermfix} for the inheriting class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
