% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermindpost.R
\name{getPostperm,mcmcpermindpost-method}
\alias{getPostperm,mcmcpermindpost-method}
\title{Getter method of \code{mcmcpermindpost} class.}
\usage{
\S4method{getPostperm}{mcmcpermindpost}(object)
}
\arguments{
\item{object}{An \code{mcmcpermindpost} object.}
}
\value{
The \code{postperm} slot of the \code{object}.
}
\description{
Returns the \code{postperm} slot.
}
\examples{
\dontrun{getPostperm(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputpermpost} for the inheriting class
\item \linkS4class{mcmcoutputpermhierpost} for the inheriting class with
hierarchical prior
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
