% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmodelmoments.R
\name{getR,normalmodelmoments-method}
\alias{getR,normalmodelmoments-method}
\title{Getter method of \code{normalmodelmoments} class.}
\usage{
\S4method{getR}{normalmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normalmodelmoments} object.}
}
\value{
The \code{R} slot of the \code{object}.
}
\description{
Returns the \code{R} slot.
}
\examples{
f_model         <- model("normal", weight = matrix(c(.3, .7), nrow = 1))
means           <- c(-2, 2)
sigmas          <- matrix(c(2, 4), nrow=1)
setPar(f_model) <- list(mu = means, sigma = sigmas)
f_moments       <- modelmoments(f_model)
getR(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
