% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studmultmodelmoments.R
\name{getRdet,studmultmodelmoments-method}
\alias{getRdet,studmultmodelmoments-method}
\title{Getter method of \code{studmultmodelmoments} class.}
\usage{
\S4method{getRdet}{studmultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studmultmodelmoments} object.}
}
\value{
The \code{Rdet} slot of the \code{object}.
}
\description{
Returns the \code{Rdet} slot.
}
\examples{
f_model         <- model("studmult", weight = matrix(c(.3, .7), nrow = 1))
means           <- matrix(c(-2, -2, 2, 2),nrow = 2)
covar           <- matrix(c(1, 1.2, 1.2, 4), nrow = 2)
sigmas          <- array(c(covar, 2*covar), dim = c(2, 2, 2))
setPar(f_model) <- list(mu = means, sigma = sigmas, df = c(10, 20))
f_moments       <- modelmoments(f_model)
getRdet(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
