% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getRowExp}
\alias{getRowExp}
\title{Getter for the \code{exp} slot in row format}
\usage{
getRowExp(object)
}
\description{
Getter for the \code{exp} slot in row format
}
\keyword{internal}
