% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{getStartpar,mcmc-method}
\alias{getStartpar,mcmc-method}
\title{Getter method of \code{mcmc} class.}
\usage{
\S4method{getStartpar}{mcmc}(object)
}
\arguments{
\item{object}{An \code{mcmc} object.}
}
\value{
The \code{startpar} slot of the \code{object}.
}
\description{
Returns the \code{startpar} slot.
}
\examples{
# Generate an mcmc object
f_mcmc <- mcmc()
# Get the slot
getStartpar(f_mcmc)

}
\seealso{
\itemize{
\item \linkS4class{mcmc} for the class definition
\item \code{\link[=mcmc]{mcmc()}} for the constructor of the \code{mcmc} class
}
}
\keyword{internal}
