% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmoments.R
\name{getVar,modelmoments-method}
\alias{getVar,modelmoments-method}
\title{Getter method of \code{modelmoments} class.}
\usage{
\S4method{getVar}{modelmoments}(object)
}
\arguments{
\item{object}{A \code{modelmoments} object.}
}
\value{
The \code{var} slot of the \code{object}.
}
\description{
Returns the \code{var} slot of a \code{modelmoments} object.
}
\examples{
f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getVar(f_moments)

}
\seealso{
\linkS4class{modelmoments} for all slots of the \code{modelmoments} class
}
\keyword{internal}
