% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentmodelmoments.R
\name{getW,studentmodelmoments-method}
\alias{getW,studentmodelmoments-method}
\title{Getter method of \code{studentmodelmoments} class.}
\usage{
\S4method{getW}{studentmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studentmodelmoments} object.}
}
\value{
The \code{W} slot of the \code{object}.
}
\description{
Returns the \code{W} slot.
}
\examples{
f_model         <- model("normal", weight = matrix(c(.3, .7), nrow = 1))
means           <- c(-2, 2)
sigmas          <- matrix(c(2, 4), nrow=1)
setPar(f_model) <- list(mu = means, sigma = sigmas, df = c(20, 40))
f_moments       <- modelmoments(f_model)
getW(f_moments)

}
\seealso{
\itemize{
\item \link{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class family
}
}
\keyword{internal}
