% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermind.R
\name{getWeightperm,mcmcpermind-method}
\alias{getWeightperm,mcmcpermind-method}
\title{Getter method of \code{mcmcpermind} class.}
\usage{
\S4method{getWeightperm}{mcmcpermind}(object)
}
\arguments{
\item{object}{An \code{mcmcpermind} object.}
}
\value{
The \code{weightperm} slot of the \code{object}.
}
\description{
Returns the \code{weightperm} slot.
}
\examples{
\dontrun{getWeightperm(mcmcperm)}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputpermbase} for the inheriting class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for function permuting MCMC samples
}
}
\keyword{internal}
