% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{getWeightperm}
\alias{getWeightperm}
\title{Getter for the \code{weightperm} slot}
\usage{
getWeightperm(object)
}
\description{
Getter for the \code{weightperm} slot
}
\keyword{internal}
