% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{getY,fdata-method}
\alias{getY,fdata-method}
\title{Getter method of \code{fdata} class.}
\usage{
\S4method{getY}{fdata}(object)
}
\arguments{
\item{object}{An \code{fdata} object.}
}
\value{
The \code{y} slot of the \code{object} in the order defined \code{bycolumn}.
}
\description{
Returns the \code{y} slot in the order defined by the slot \code{bycolumn}.
}
\examples{
# Create an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
getY(f_data)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
