% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphic_func.R
\name{graphic_func}
\alias{graphic_func}
\alias{.check.grDevice}
\title{Checks if graphical device has \code{title} option}
\usage{
.check.grDevice()
}
\value{
\code{TRUE} if \code{title} option exists.
}
\description{
For internal use only.
}
\keyword{internal}
