% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmoments.R
\docType{class}
\name{groupmoments-class}
\alias{groupmoments-class}
\alias{.groupmoments}
\alias{show,groupmoments-method}
\title{Finmix \code{groupmoments} class}
\usage{
\S4method{show}{groupmoments}(object)
}
\arguments{
\item{object}{A \code{groupmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Stores moments for finite mixture component distributions. These are only
available, if the data contains in addition to observations also indicators
defining to which component a certain observation belongs. These indicators
are stored in an \linkS4class{fdata} object in the slot \code{S}.

Calling \code{\link[=show]{show()}} on a \code{groupmoments} object gives an overview
of the moments of a finit mixture with continuous data.
}
\section{Functions}{
\itemize{
\item \code{show,groupmoments-method}: Shows a short summary of the object's slots
}}

\section{Slots}{

\describe{
\item{\code{NK}}{An array containing the group sizes for each component.}

\item{\code{mean}}{A matrix containing the group averages for each component.}

\item{\code{WK}}{An array containing the within-group variability. For multivariate
data this is an array of dimension \verb{K x r x r} and for univariate
data this is simply an array of dimension \verb{1 x K}.}

\item{\code{var}}{An array containing the within-group (co)variance. For multivariate
data this is an array of dimension \verb{K x r x r} and for univariate
data this is simply an array of dimension \verb{1 x K}.}

\item{\code{fdata}}{An \linkS4class{fdata} object containing the data.}
}}

\seealso{
\itemize{
\item \code{\link[=groupmoments]{groupmoments()}} for the class constructor
\item \linkS4class{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
}
}
\keyword{internal}
