% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmoments.R
\name{groupmoments}
\alias{groupmoments}
\title{Finmix \code{groupmoments} class constructor}
\usage{
groupmoments(value = fdata())
}
\arguments{
\item{value}{An \code{fdata} object containing observations in slot \code{y} and
indicators in slot \code{S}.}
}
\value{
A \code{groupmoments} object containing component-specific moments of the
\code{fdata} object.
}
\description{
Calling \code{\link[=groupmoments]{groupmoments()}} creates an object holding various
component-specific moments. These moments can only constructed if the
\link[=fdata_class]{fdata} object contains in addition to observations also
indicators defining from which component a certain observation stems.
}
\examples{
# Define a mixture model with exponential components.
f_model <- model("exponential", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Create group moments of the data.
groupmoments(f_data)

}
\seealso{
\itemize{
\item \link[=fdata_class]{fdata} for the \code{fdata} class definition
\item \link[=groupmments_class]{groupmoments} for the definition of the \code{groupmoments}
class
\item \link[=datamoments_class]{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
}
}
