% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{hasExp,fdata-method}
\alias{hasExp,fdata-method}
\title{Checker method for \code{exp} slot of an \code{fdata} object.}
\usage{
\S4method{hasExp}{fdata}(object, verbose = FALSE)
}
\arguments{
\item{object}{An \code{fdata} object.}

\item{verbose}{A logical indicating, if the function should print out
messages.}
}
\value{
Either \code{FALSE}/\code{TRUE}, if \code{verbose} is \code{FALSE} and the \code{exp} slot is
empty or filled or a message, if \code{verbose} is \code{TRUE}.
}
\description{
\code{hasExp()} checks, if the object contains \code{exp} data.
}
\examples{
# Generate an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
hasExp(f_data)

}
\seealso{
\itemize{
\item \linkS4class{fdata} for the class definition
}
}
\keyword{internal}
