% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasExp}
\alias{hasExp}
\title{Checks for the \code{exp} slot of an \code{fdata} object}
\usage{
hasExp(object, verbose = FALSE)
}
\description{
Checks for the \code{exp} slot of an \code{fdata} object
}
\keyword{internal}
