% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasPar,model-method}
\alias{hasPar,model-method}
\title{Checks for parameters.}
\usage{
\S4method{hasPar}{model}(object, verbose = FALSE)
}
\arguments{
\item{object}{An S4 model object.}

\item{verbose}{A logical indicating, if the function should give a print out.}
}
\value{
A matrix with repetitions. Can be empty, if no repetitions are set.
}
\description{
\code{hasPar} checks if the model has parameters defined.
}
\examples{
\dontrun{
if(hasPar(model)) {simulate(model)}
}

}
\seealso{
\code{model}
}
