% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasPar}
\alias{hasPar}
\title{Checks a finite mixture model for the parameters}
\usage{
hasPar(object, verbose = FALSE)
}
\description{
Checks a finite mixture model for the parameters
}
\keyword{internal}
