% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{hasPriorPar,prior,model-method}
\alias{hasPriorPar,prior,model-method}
\title{Checks for parameters in a \code{prior} object}
\usage{
\S4method{hasPriorPar}{prior,model}(object, model, verbose = FALSE)
}
\arguments{
\item{object}{A \code{prior} object containing the specifications for the prior.}

\item{model}{A \code{model} object containing the specifications for the model.}

\item{verbose}{A logical indicating, if the output should be verbose.}
}
\description{
Calling \code{hasPriorPar()} checks if \code{model}-appropriate parameters are stored
in the \code{prior} object.
}
\examples{
# Define a Poisson mixture model.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Call the default constructor.
f_prior <- prior()
# Check if the prior has appropriate parameters defined.
hasPriorPar(f_prior, f_model)
\dontrun{hasPriorPar(f_prior, f_model, TRUE)}

}
\seealso{
\itemize{
\item \linkS4class{prior} for the definition of the \code{prior} class
\item \linkS4class{model} for the definition of the \code{model} class
}
}
\keyword{internal}
