% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasPriorPar}
\alias{hasPriorPar}
\title{Checks for the \code{par} slot in the \code{prior} class}
\usage{
hasPriorPar(object, model, verbose = FALSE)
}
\description{
Checks for the \code{par} slot in the \code{prior} class
}
\keyword{internal}
