% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{hasS,fdata-method}
\alias{hasS,fdata-method}
\title{Checker method for \code{S} slot of an \code{fdata} object.}
\usage{
\S4method{hasS}{fdata}(object, verbose = FALSE)
}
\arguments{
\item{object}{An \code{fdata} object.}

\item{verbose}{A logical indicating, if the function should print out
messages.}
}
\value{
Either \code{FALSE}/\code{TRUE}, if \code{verbose} is \code{FALSE} and the \code{S} slot is
empty or filled or a message, if \code{verbose} is \code{TRUE}.
}
\description{
\code{\link[=hasY]{hasY()}} checks, if the object contains \code{S} data.
}
\examples{
# Generate an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
hasS(f_data)

}
\seealso{
\link{fdata} class for an overview of its slots
}
