% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasS}
\alias{hasS}
\title{Checks for the \code{S} slot of an \code{fdata} object}
\usage{
hasS(object, verbose = FALSE)
}
\description{
Checks for the \code{S} slot of an \code{fdata} object
}
\keyword{internal}
