% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasT,model-method}
\alias{hasT,model-method}
\title{Checks for repetitions.}
\usage{
\S4method{hasT}{model}(object, verbose = FALSE)
}
\arguments{
\item{verbose}{A logical indicating if the function should give a print out.}

\item{model}{An S4 model object.}
}
\value{
A logical. \code{TRUE} if repetitions are existent in the model. If
values of slot \code{T} are \code{NA} it returns \code{FALSE}.
}
\description{
\code{hasT} chwecks if the model object possesses repetitions.
}
\examples{
\dontrun{
if(hasT(model)) {cat('Has repetitions.')}
}

}
\seealso{
\code{model}
}
\keyword{internal}
