% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasT}
\alias{hasT}
\title{Checks a finite mixture model for repetitions}
\usage{
hasT(object, verbose = FALSE)

hasT(object, verbose = FALSE)
}
\description{
Checks a finite mixture model for repetitions

Checks for the \code{T} slot of an \code{fdata} object
}
\keyword{internal}
