% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasWeight,model-method}
\alias{hasWeight,model-method}
\title{Getter for weights}
\usage{
\S4method{hasWeight}{model}(object, verbose = FALSE)
}
\arguments{
\item{verbose}{A logical indicating, if the function should give a print out.}

\item{model}{An S4 model object.}
}
\value{
Matrix of weights.
}
\description{
\code{hasWeight} returns the weight matrix.
}
\examples{
\dontrun{
weight <- hasWeight(model)
}
}
\keyword{internal}
