% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasWeight}
\alias{hasWeight}
\title{Checks a finite mixture model for the weight}
\usage{
hasWeight(object, verbose = FALSE)
}
\description{
Checks a finite mixture model for the weight
}
\keyword{internal}
