% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{hasY,fdata-method}
\alias{hasY,fdata-method}
\title{Checker method for \code{y} slot of an \code{fdata} object.}
\usage{
\S4method{hasY}{fdata}(object, verbose = FALSE)
}
\arguments{
\item{object}{An \code{fdata} object.}

\item{verbose}{A logical indicating, if the function should print out
messages.}
}
\value{
Either \code{FALSE}/\code{TRUE}, if \code{verbose} is \code{FALSE} and the \code{y} slot is
empty or filled or a message, if \code{verbose} is \code{TRUE}.
}
\description{
\code{hasY()} checks, if the object contains \code{y} data.
}
\examples{
# Generate an fdata object with Poisson data
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
hasY(f_data)

}
\seealso{
\link{fdata} class for an overview of its slots
}
\keyword{internal}
