% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{hasY}
\alias{hasY}
\title{Checks for the \code{y} slot of an \code{fdata} object}
\usage{
hasY(object, verbose = FALSE)
}
\description{
Checks for the \code{y} slot of an \code{fdata} object
}
\keyword{internal}
