% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hungarian_cc}
\alias{hungarian_cc}
\title{Compute the hungarian matrix}
\usage{
hungarian_cc(cost)
}
\arguments{
\item{cost}{A matrix containing the costs for each row source and column
target.}
}
\value{
An integer matrix defining the best solution to the assignment
problem.
}
\description{
This function calls an implementation of the Hungarian algorithm by Munkres.
The Hungarian algorithm solves a weighted assignment problem on a bipartite
graph. Note, here this algorithm is used in the re-labeling algorithm by
Stephens (1997b).
}
\references{
\itemize{
\item Stephens, Matthew (1997b), "Dealing with Label-Switching in Mixture
Models", Journal of the Royal Statistical Society Series B, 62(4)
}
}
\seealso{
\itemize{
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for the calling function
\item \code{\link[=mcmcestimate]{mcmcestimate()}} for the function that uses the re-labeling algorithm by
Stephens (1997b)
}
}
