% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdatamoments.R
\name{initialize,cdatamoments-method}
\alias{initialize,cdatamoments-method}
\title{Initializer of the \code{cdatamoments} class}
\usage{
\S4method{initialize}{cdatamoments}(.Object, ..., value = fdata())
}
\arguments{
\item{.Object}{An object: see the "initialize Methods" section in
\link{initialize}.}

\item{...}{Arguments to specify properties of the new object, to be passed
to \code{initialize()}.}

\item{model}{A finmix \code{fdata} object containing the observations.}
}
\description{
Only used implicitly. The initializer calls a function \code{generateMoments()}
to generate in the initialization step the moments for a passed-in \code{fdata}
object.
}
\seealso{
\itemize{
\item \link{Classes_Details} for details of class definitions, and
\item \link{setOldClass} for the relation to S3 classes
}
}
\keyword{internal}
