% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixhier.R
\name{initialize,mcmcoutputpermfixhier-method}
\alias{initialize,mcmcoutputpermfixhier-method}
\title{Initializer of the \code{mcmcoutputpermfixhier} class}
\usage{
\S4method{initialize}{mcmcoutputpermfixhier}(
  .Object,
  mcmcoutput,
  Mperm = integer(),
  parperm = list(),
  logperm = list(),
  hyperperm = list()
)
}
\arguments{
\item{.Object}{An object: see the "initialize Methods" section in
\link{initialize}.}

\item{mcmcoutput}{An \code{mcmcoutput} class containing the results from MCMC
sampling.}

\item{Mperm}{An integer defining the number of permuted MCMC samples.}

\item{parperm}{A named list containing the permuted component parameter
samples from MCMC sampling}

\item{logperm}{A named list containing the mixture log-likelihood, the
prior log-likelihood, and the complete data posterior log-likelihood
for the permuted MCMC samples.}

\item{hyperperm}{A named list containing the permuted parameters of the
hierarchical prior.}
}
\description{
Only used implicitly. The initializer stores the data into the slots of the
passed-in object.
}
\seealso{
\itemize{
\item \link{Classes_Details} for details of class definitions, and
\item \link{setOldClass} for the relation to S3 classes
}
}
\keyword{internal}
