% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhierpost.R
\name{initialize,mcmcoutputpermhierpost-method}
\alias{initialize,mcmcoutputpermhierpost-method}
\title{Initializer of the \code{mcmcoutputpermhierpost} class}
\usage{
\S4method{initialize}{mcmcoutputpermhierpost}(
  .Object,
  mcmcoutput,
  Mperm = integer(),
  parperm = list(),
  relabel = character(),
  weightperm = array(),
  logperm = list(),
  hyperperm = list(),
  postperm = list(),
  entropyperm = array(),
  STperm = array(),
  Sperm = array(),
  NKperm = array()
)
}
\arguments{
\item{.Object}{An object: see the "initialize Methods" section in
\link{initialize}.}

\item{mcmcoutput}{An \code{mcmcoutput} class containing the results from MCMC
sampling.}

\item{Mperm}{An integer defining the number of permuted MCMC samples.}

\item{parperm}{A named list containing the permuted component parameter
samples from MCMC sampling}

\item{relabel}{A character specifying the relabeling algorithm used for
permuting the MCMC samples.}

\item{weightperm}{An array of dimension \verb{Mperm x K} containing the
relabeled weight parameters.}

\item{logperm}{A named list containing the mixture log-likelihood, the
prior log-likelihood, and the complete data posterior log-likelihood
for the permuted MCMC samples.}

\item{hyperperm}{A named list containing the (permuted) parameters of the
hierarchical prior.}

\item{postperm}{A named list containing a named list \code{par} with array(s) of
parameters from the posterior density.}

\item{entropyperm}{An \code{array} of dimension \verb{Mperm x 1} containing the
entropy for each MCMC permuted draw.}

\item{STperm}{An \code{array} of dimension \verb{Mperm x 1} containing all permuted
MCMC states, for the last observation in slot \verb{@y} of the \code{fdata} object
passed in to \code{\link[=mixturemcmc]{mixturemcmc()}} where a state is defined for non-Markov
models as the last indicator of this observation.}

\item{NKperm}{An \code{array} of dimension \verb{Mperm x K} containing the numbers
of observations assigned to each component.}

\item{An}{\code{array} of dimension \verb{N x storeS} containing the last
\code{storeS} permuted indicators. \code{storeS} is defined in the slot \verb{@storeS}
of the \code{mcmc} object passed into \code{\link[=mixturemcmc]{mixturemcmc()}}.}
}
\description{
Only used implicitly. The initializer stores the data into the slots of the
passed-in object.
}
\seealso{
\itemize{
\item \link{Classes_Details} for details of class definitions, and
\item \link{setOldClass} for the relation to S3 classes
}
}
\keyword{internal}
