% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonmodelmoments.R
\name{initialize,poissonmodelmoments-method}
\alias{initialize,poissonmodelmoments-method}
\title{Initializer of the \code{poissonmoments} class}
\usage{
\S4method{initialize}{poissonmodelmoments}(.Object, ..., model)
}
\arguments{
\item{.Object}{An object_ see the "initialize Methods" section in
\link{initialize}.}

\item{...}{Arguments to specify properties of the new object, to be passed
to \code{initialize()}.}

\item{model}{A finmix \code{model} object containing the definition of the
finite mixture distribution.}
}
\description{
Only used implicitly. The initializer calls a function \code{generateMoments()}
to generate in the initialization step also the moments for a passed \code{model}
object.
}
\seealso{
\itemize{
\item \link{Classes_Details} for details of class definitions, and
\item \link{setOldClass} for the relation to S3 classes
}
}
\keyword{internal}
