% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lddirichlet_cc}
\alias{lddirichlet_cc}
\title{Computes the log density of the Dirichlet distribution}
\usage{
lddirichlet_cc(values, par)
}
\arguments{
\item{values}{A matrix of dimension \verb{M x K} for which the log-density
should be calculated.}

\item{par}{A vector of dimension \verb{K x 1} containing the Dirichlet
parameters.}
}
\value{
A vector of Dirichlet log-density values.
}
\description{
For each shape and rate parameter pair the log-Dirichlet density is
computed. Inside the function the unsafe access functions of Armadillo
\code{at()} and \code{unsafe_col()} are used, so now boundary check is performed.
In each step the \code{lgammafn()} function from Rcpp's \code{R} namespace is used.
At this time unused.
}
