% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\docType{class}
\name{mcmc-class}
\alias{mcmc-class}
\alias{.mcmc}
\title{Finmix \code{mcmc} class}
\description{
This class defines hyper-parameters for the MCMC procedure. This is a main
class of the \code{finmix} package that must be defined for estimating a finite
mixture model.
}
\section{Slots}{

\describe{
\item{\code{burnin}}{An integer defining the number of steps in the burn-in phase of
Gibbs-sampling.}

\item{\code{M}}{An integer defining the number of steps in Gibbs-sampling to be
stored.}

\item{\code{startpar}}{A logical indicating, if starting by sampling the
parameters. If \code{FALSE} sampling starts by sampling the indicators \code{S}.}

\item{\code{storeS}}{An integer specifying how many of the last sampled indicators
should be stored in the output.}

\item{\code{storepost}}{A logical indicating if the posterior probabilities should
be stored. This becomes for example important for specific relabeling
algorithms, but also for analysis.}

\item{\code{ranperm}}{A logical indicating, if random permutation should be used. If
\code{TRUE} the parameters are permutated randomly between the number of
components after each sampling step in MCMC.}

\item{\code{storeinv}}{A logical indicating if the inverse variance-covariance
matrices for multivariate normal or Student-t mixtures should be stored.}
}}

\seealso{
\itemize{
\item \code{\link[=mcmc]{mcmc()}} for the class constructor
\item \code{\link[=mcmcstart]{mcmcstart()}} for completion of slots
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for further information about the MCMC sampling
}
}
