% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{mcmc}
\alias{mcmc}
\title{Constructor for \code{mcmc} class}
\usage{
mcmc(
  burnin = 0,
  M = 5000,
  startpar = TRUE,
  storeS = 1000,
  storepost = TRUE,
  ranperm = TRUE,
  storeinv = TRUE
)
}
\arguments{
\item{burnin}{An integer defining the number of steps in the burn-in phase of
Gibbs-sampling.}

\item{M}{An integer defining the number of steps in Gibbs-sampling to be
stored.}

\item{startpar}{A logical indicating, if starting by sampling the
parameters. If \code{FALSE} sampling starts by sampling the indicators \code{S}.}

\item{storeS}{An integer specifying how many of the last sampled indicators
should be stored in the output.}

\item{storepost}{A logical indicating if the posterior probabilities should
be stored. This becomes for example important for specific relabeling
algorithms, but also for analysis.}

\item{ranperm}{A logical indicating, if random permutation should be used. If
\code{TRUE} the parameters are permutated randomly between the number of
components after each sampling step in MCMC.}

\item{storeinv}{A logical indicating if the inverse variance-covariance
matrices for multivariate normal or Student-t mixtures should be stored.}
}
\value{
An object of class \code{mcmc} containing all hyper-parameters for MCMC
sampling.
}
\description{
Calling \code{\link[=mcmc]{mcmc()}} constructs an object of class \code{mcmc} that specifies the
hyper-parameters for the MCMC procedure. Each MCMC sampling needs an \code{mcmc}
object that specifies the way, how MCMC sampling should be performed and
what kind and how much of data should be stored.
}
\examples{
f_mcmc <- mcmc()

}
\seealso{
\itemize{
\item \link[=mcmc_class]{mcmc} for the definition of the \code{mcmc} class
\item \code{\link[=mcmcstart]{mcmcstart()}} for setting up all objects for MCMC sampling
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for running MCMC sampling for finite mixture models
}
}
