% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_normult_cc}
\alias{mcmc_normult_cc}
\title{Performs MCMC sampling for mixtures of multivariate Normal distributions}
\usage{
mcmc_normult_cc(data_S4, model_S4, prior_S4, mcmc_S4, mcmcoutput_S4)
}
\arguments{
\item{data_S4}{An \code{fdata} object storing the observations and indicators.}

\item{model_S4}{A \code{model} object specifying the multivariate Normal finite
mixture model.}

\item{prior_S4}{A \code{prior} object specifying the prior distribution for MCMC
sampling.}

\item{mcmc_S4}{A \code{mcmc} object specifying the hyper-parameters for MCMC
sampling.}

\item{mcmcoutput_S4}{An \code{mcmcoutput} object storing the outcomes from MCMC
sampling using R memory.}
}
\value{
An \code{mcmcoutput} object containing the results from MCMC sampling
and using the R memory from the input argument \code{mcmcoutput_S4}.
}
\description{
For internal usage only. This function gets passed the \code{fdata}, \code{model},
\code{prior}, \code{mcmc} objects to perform MCMC sampling for a multivariate Normal
mixture model. In addition an \code{mcmcoutput} object is given that stores the
output of MCMC sampling in R memory. Note that \code{finmix} relies in C++ code
on so-called "mixin" layers that help to design a software by organizing
code into layers that build upon each others and enable modularity in MCMC
sampling by allowing to combine different sampling designs, e.g. with or
without a hierarchical prior, with fixed indicators or storing posterior
density parameters. See for more information on mixin layers Smaragdakis
and Butory (1998).
}
\references{
\itemize{
\item Smaragdakis, Y. and Butory, D. (1998), "Implementing layered designs with
mixin layers." In: Jul E. (eds) ECOOP’98 — Object-Oriented Programming.
ECOOP 1998. Lecture Notes in Computer Science, vol 1445. Springer,
Berlin, Heidelberg.
}
}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \link[=fdata_class]{fdata} for the \code{fdata} class definition
\item \link[=model_class]{model} for the \code{model} class definition
\item \link[=prior-class]{prior} for the \code{prior} class definition
\item \link[=mcmc_class]{mcmc} for the \code{mcmc} class definition
}
}
