% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestind.R
\docType{class}
\name{mcmcest-class}
\alias{mcmcest-class}
\title{Finmix \code{mcmcest} class}
\description{
This class stores Bayesian parameter estimates from MCMC samples and
corresponding metadata. Calling \code{\link[=mcmcestimate]{mcmcestimate()}} returns an object of this
class.
}
\details{
Calling \code{\link[=mcmcestimate]{mcmcestimate()}} on an object of class \code{mcmcoutput} or
\code{mcmcoutputperm} returns an object of class \code{mcmcest} that contains all
Bayesian estimates together with corresponding metadata. Three Bayesian
point estimates are constructed:
\itemize{
\item \strong{BML}: The Bayesian Maximum Likelihood, which is the parameter sample
from MCMC sampling that maximizes the mixture likelihood.
\item \strong{MAP}: The Bayesian Maximum A Posterior, which is the parameter sample
from MCMC sampling that maximizes the the posterior maximum likelihood.
\item \strong{EAVG}: The Ergodic Average over the MCMC samples without identification.
\item \strong{IEAVG}: The Identified Ergodic Average over the MCMC samples with
identification.
}

Note that a model with fixed indicators (i.e. slot \code{indicfix=TRUE}) has
always an identified ergodic average, because in each MCMC sample the
component labels are the same and therefore identified. In contrast, a
model with unknown indicators (i.e. \code{indicfix=FALSE}) suffers usually under
random label switching during sampling and therefore the ergodic average
over all MCMC samples is usually not identified as it averages over
parameters from different components thereby pulling the component
parameters together (sometimes you get the same average for all components).
The \code{ieavg} is calculated for a model with unknown indicators by relabeling
the component parameter samples. Re-labeling reassigns component parameters
to the most likely label of the mixture in regard to the observations. As a
result the \code{mcmcest} object of a model with unknown indicators will have
both, an \code{eavg} and an \code{ieavg} slot containing the ergodic average over
samples before and after re-labeling. The \code{relabel} slot indicates which
re-labeling algorithm had been chosen.

The uncertainty of parameter estimates is measured by the standard deviation
over parameters from MCMC sampling and stored in the \code{sdpost} slot. It is
an estimate of the standard deviation of the true posterior parameter
distribution.

The class \code{mcmcest} is a class union and includes all classes that define
objects for Bayesian estimates of MCMC samples and is used to dispatch
methods for \code{mcmcest} objects. For the user this detail is not important,
especially as this class has no exported constructor. Objects are solely
constructed internally within the function \code{\link[=mcmcestimate]{mcmcestimate()}}.
\subsection{Class Methods}{

Similar to the contained classes \link[=mcmcoutput-class]{mcmcoutput} this class comes
along with a couple of methods that should give the user some comfort in
handling the permuted sampling results. There are no setters for this class
as the slots are only set internally.
\subsection{Show and Summary}{
\itemize{
\item \code{show()} gives a short summary of the object's slots.
\item \code{Summary()} prints out a summary of estimation results.
}
}

\subsection{Getters}{
\itemize{
\item \code{getDist()} returns the \code{dist} slot.
\item \code{getK()} returns the \code{K} slot.
\item \code{getIndicmod()} returns the \code{indicmod} slot.
\item \code{getBurnin()} returns the \code{burnin} slot.
\item \code{getM()} returns the \code{M} slot.
\item \code{getRanperm()} returns the \code{ranperm} slot.
\item \code{getRelabel()} returns the \code{relabel} slot.
\item \code{getMap()} returns the MAP estimates.
\item \code{getBml()} returns the BML estimates.
\item \code{getEavg()} returns the EAVG estimates.
\item \code{getIeavg()} returns the identified EAVG estimates.
\item \code{getSdpost()} returns the \code{sdpost}.
}
}

}
}
\section{Slots}{

\describe{
\item{\code{dist}}{A character specifying the distribution family of the mixture
model used in MCMC sampling.}

\item{\code{K}}{An integer specifying the number of components in the mixture model.}

\item{\code{indicmod}}{A character specifying the indicator model. At this moment
only a multinomial model can be chosen.}

\item{\code{burnin}}{An integer specifying the number of iterations in the burn-in
phase of MCMC sampling.}

\item{\code{M}}{An integer specifying the number of iterations to store in MCMC
sampling.}

\item{\code{ranperm}}{A logical specifying, if random permutation has been used
during MCMC sampling.}

\item{\code{relabel}}{A character specifying the re-labeling algorithm used during
parameter estimation for the identified ergodic average.}

\item{\code{map}}{A named list containing the parameter estimates of the MAP. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{bml}}{A named list containing the parameter estimates of the BML. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{eavg}}{A named list containing the parameter estimates of the
unidentified EAVG. Note that this is only the case for a model with
unknown indicators.}

\item{\code{ieavg}}{A named list containing the parameter estimates of the IEAVG. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{sdpost}}{A named list containing the standard deviations of the
parameter estimates from the posterior distributions.}
}}

