% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestfix.R, R/mcmcestind.R
\docType{class}
\name{mcmcest_class}
\alias{mcmcest_class}
\alias{.mcmcestfix}
\alias{mcmcestind-class}
\alias{.mcmcestind}
\title{Finmix \code{mcmcestfix} class}
\description{
This class stores the point estimators for component parameters and weights
as well as corresponding information from MCMC sampling. Three point
estimators are calculated: the maximum a posterior (MAP), the Bayesian
maximum likelihood (BML) and the Identified ergodic average (IEAVG). See
Fr\"uhwirth-Schnatter (2006) for detailed information about how these
estimators are defined.

This class stores the point estimators for component parameters and weights
as well as corresponding information from MCMC sampling. Three point
estimators are calculated: the maximum a posterior (MAP), the Bayesian
maximum likelihood (BML) and the Identified ergodic average (IEAVG). See
Fr\"uhwirth-Schnatter (2006) for detailed information about how these
estimators are defined.

Note that this class inherits almost all of its slots from the \code{mcmcestfix}
class, the corresponding class for fixed indicators.
}
\section{Functions}{
\itemize{
\item \code{mcmcestind-class}: Finmix \code{mcmcestind} class
}}

\section{Slots}{

\describe{
\item{\code{dist}}{A character specifying the distribution family of the mixture
model used in MCMC sampling.}

\item{\code{K}}{An integer specifying the number of components in the mixture model.}

\item{\code{indicmod}}{A character specifying the indicator model. At this moment
only a multinomial model can be chosen.}

\item{\code{burnin}}{An integer specifying the number of iterations in the burn-in
phase of MCMC sampling.}

\item{\code{M}}{An integer specifying the number of iterations to store in MCMC
sampling.}

\item{\code{ranperm}}{A logical specifying, if random permutation has been used
during MCMC sampling.}

\item{\code{relabel}}{A character specifying the re-labeling algorithm used during
parameter estimation for the identified ergodic average.}

\item{\code{map}}{A named list containing the parameter estimates of the MAP. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{bml}}{A named list containing the parameter estimates of the BML. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{A}}{named list containing the parameter estimates of the IEAVG. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{sdpost}}{A named list containing the standard deviations of the
parameter estimates from the posterior distributions.}

\item{\code{eavg}}{A named list containing the estimates of the ergodic average. The
element \code{par} is a list and contains the component parameter estimates and
\code{weight} contains the weight estimates. The difference between the EAVG
and the IEAVG is that the IEAVG is based on re-labeled samples.}
}}

\seealso{
\itemize{
\item \link[=mcmcest_class]{mcmcestind} for the equivalent class for models with
unknown indicators
\item \code{\link[=mcmcestimate]{mcmcestimate()}} to calculate point estimates
}

\itemize{
\item \link[=mcmcest_class]{mcmcestfix} for the parent class with fixed indicators
\item \code{\link[=mcmcestimate]{mcmcestimate()}} to calculate point estimates
}
}
