% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestfix.R
\docType{class}
\name{mcmcestfix-class}
\alias{mcmcestfix-class}
\alias{.mcmcestfix}
\title{Finmix \code{mcmcestfix} class}
\description{
This class stores the point estimators for component parameters and weights
as well as corresponding information from MCMC sampling. Three point
estimators are calculated: the maximum a posterior (MAP), the Bayesian
maximum likelihood (BML) and the Identified ergodic average (IEAVG). See
Fr\"uhwirth-Schnatter (2006) for detailed information about how these
estimators are defined.
}
\section{Slots}{

\describe{
\item{\code{dist}}{A character specifying the distribution family of the mixture
model used in MCMC sampling.}

\item{\code{K}}{An integer specifying the number of components in the mixture model.}

\item{\code{indicmod}}{A character specifying the indicator model. At this moment
only a multinomial model can be chosen.}

\item{\code{burnin}}{An integer specifying the number of iterations in the burn-in
phase of MCMC sampling.}

\item{\code{M}}{An integer specifying the number of iterations to store in MCMC
sampling.}

\item{\code{ranperm}}{A logical specifying, if random permutation has been used
during MCMC sampling.}

\item{\code{relabel}}{A character specifying the re-labeling algorithm used during
parameter estimation for the identified ergodic average.}

\item{\code{map}}{A named list containing the parameter estimates of the MAP. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{bml}}{A named list containing the parameter estimates of the BML. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{ieavg}}{A named list containing the parameter estimates of the IEAVG. The
element \code{par} is a named list and contains the component parameters and
the element \code{weight} contains the weights.}

\item{\code{sdpost}}{A named list containing the standard deviations of the
parameter estimates from the posterior distributions.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcestind} for the equivalent class for models with
unknown indicators
\item \code{\link[=mcmcestimate]{mcmcestimate()}} to calculate point estimates
}
}
\keyword{internal}
