% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestimate.R
\name{mcmcestimate}
\alias{mcmcestimate}
\title{Calculate point estimators from MCMC samples}
\usage{
mcmcestimate(
  mcmcout,
  method = "kmeans",
  fdata = NULL,
  permOut = FALSE,
  opt_ctrl = list(max_iter = 200L)
)
}
\arguments{
\item{mcmcout}{An \code{mcmcoutput} object containing the sampled parameters and
informaiton about the finite mixture model.}

\item{method}{A character defining the re-labeling method in case of a model
with unknown indicators. For most distributions there exists only a single
choice, namely "kmeans". For Poisson and Binomial distributions the
re-labeling algorithms "Stephens1997a" and "Stephens1997b" can be chosen.}

\item{fdata}{An \code{fdata} model containing the observations. Optional.}

\item{permOut}{A logical indicating, if the permuted MCMC samples should be
returned as well. Optional.}

\item{opt_ctrl}{A list with an element \code{max_iter} controlling the number of
iterations in case the "Stephens1997a" re-labeling algorithm is chosen.}
}
\value{
An \code{mcmcest} object cotnaining the point estimates together with
additional information about the underlying finite mixture model, MCMC
sampling hyper-parameters and the data. In case \code{permOut} is set to
\code{TRUE}, the output of this function is a named list with an \code{mcmcest}
object containing parameter estimates and in addition an \code{mcmcoutputperm}
object containing the permuted (re-labeled) MCMC samples.
}
\description{
Calling \code{\link[=mcmcestimate]{mcmcestimate()}} calculates the following point estimates from the
MCMC samples:
\itemize{
\item MAP: The maximum a posterior estimates are defined as the mode of the
(joint) posterior density.
\item BML: The Bayesian maximum likelihood estimator is based on the mixture
log-likelihood function and defines the mode of this function.
\item EAVG: The ergodic average is calculated as an average over the MCMC traces
of component parameters and weights (in case of unknown parameters).
\item IEAVG: The identified ergodic average is defined similar to the EAVG,
however, in contrast to the latter it is based on re-labeled MCMC traces.
This is especially important in case of random permutation during MCMC
sampling as component parameters then have to be re-assigned to their
(probably) correct component.
}

For a more detailed outlay of point estimators from Bayesian mixture model
estimation, see Fr\"uhwirth-Schnatter (2006).
}
\seealso{
\itemize{
\item \link[=mcmcest_class]{mcmcestfix} for object storing the parameter estimates in
case of fixed indicators
\item \link[=mcmcest_class]{mcmcestfix} for object storing the parameter estimates in
case of unknown indicators
\item \linkS4class{mcmcoutputperm} for classes storing re-labeled
MCMC samples
}
}
