% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestind.R
\docType{class}
\name{mcmcestind-class}
\alias{mcmcestind-class}
\alias{.mcmcestind}
\title{Finmix \code{mcmcestfix} class}
\description{
This class stores the point estimators for component parameters and weights
as well as corresponding information from MCMC sampling. Three point
estimators are calculated: the maximum a posterior (MAP), the Bayesian
maximum likelihood (BML) and the Identified ergodic average (IEAVG). See
Fr\"uhwirth-Schnatter (2006) for detailed information about how these
estimators are defined.

Note that this class inherits almost all of its slots from the \code{mcmcestfix}
class, the corresponding class for fixed indicators.
}
\section{Slots}{

\describe{
\item{\code{eavg}}{A named list containing the estimates of the ergodic average. The
element \code{par} is a list and contains the component parameter estimates and
\code{weight} contains the weight estimates. The difference between the EAVG
and the IEAVG is that the IEAVG is based on re-labeled samples.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcestfix} for the parent class with fixed indicators
\item \code{\link[=mcmcestimate]{mcmcestimate()}} to calculate point estimates
}
}
\keyword{internal}
