% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcextract.R
\docType{class}
\name{mcmcextract-class}
\alias{mcmcextract-class}
\alias{.mcmcextract}
\title{Finmix \code{mcmcextract} class}
\description{
This is a leight-weighted class containing the major results from MCMC
sampling to calculate model moments from MCMC samples. Note that momentarily
only methods for the multivariate Normal mixture are implemented.
}
\section{Slots}{

\describe{
\item{\code{dist}}{A character defining the finite mixture model that has been used
in MCMC sampling.}

\item{\code{K}}{An integer specifying the number of components of the mixture model.}

\item{\code{r}}{An integer specifying the number of dimensions of the mixture model.}

\item{\code{par}}{A list storing the sample component parameters from MCMC sampling.}

\item{\code{weight}}{A n array storing the sample weight parameters from MCMC
sampling.}
}}

\keyword{internal}
