% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhierpost.R
\docType{class}
\name{mcmcoutput-class}
\alias{mcmcoutput-class}
\title{Finmix \code{mcmcoutput} class union}
\description{
This class union is set to dispatch methods for \code{mcmcoutput} objects from
MCMC sampling.
}
