% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfix.R
\docType{class}
\name{mcmcoutputfix-class}
\alias{mcmcoutputfix-class}
\alias{.mcmcoutputfix}
\title{Finmix \code{mcmcoutputfix} class}
\description{
This class defines the basic slots for the MCMC sampling output for a
fixed indicator model.
}
\section{Slots}{

\describe{
\item{\code{M}}{An integer defining the number of iterations in MCMC sampling.}

\item{\code{burnin}}{An integer defining the number of iterations in the burn-in
phase of MCMC sampling. These number of sampling steps are not stored
in the output.}

\item{\code{ranperm}}{A logical indicating, if MCMC sampling has been performed
with random permutations of components.}

\item{\code{par}}{A named list containing the sampled component parameters.}

\item{\code{log}}{A named list containing the values of the mixture log-likelihood,
mixture prior log-likelihood, and the complete data posterior
log-likelihood.}

\item{\code{model}}{The \code{model} object that specifies the finite mixture model for
whcih MCMC sampling has been performed.}

\item{\code{prior}}{The \code{prior} object defining the prior distributions for the
component parameters that has been used in MCMC sampling.}
}}

\keyword{internal}
