% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhier.R
\docType{class}
\name{mcmcoutputfixhier-class}
\alias{mcmcoutputfixhier-class}
\alias{.mcmcoutputfixhier}
\title{Finmix \code{mcmcoutput} class for hierarchical priors}
\description{
This class stores in addition to the information from its parent class
\code{mcmcoutputfix} also the sampled parameters from the hierarchical prior.
}
\section{Slots}{

\describe{
\item{\code{hyper}}{A list storing the sampled parameters from the hierarchical
prior.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputfix} for the parent class``
}
}
