% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhierpost.R
\docType{class}
\name{mcmcoutputfixhierpost-class}
\alias{mcmcoutputfixhierpost-class}
\alias{.mcmcoutputfixhierpost}
\title{Finmix \code{mcmcoutputfixhierpost} class}
\description{
This class inherits from the \code{mcmcoutputfixhier} class and adds posterior
density parameters to the MCMC sampling output. The storage of posterior
parameters is controlled by the slot \code{storepost} in the \linkS4class{mcmc}
class. If set to \code{TRUE} posterior parameters are stored in the output of the
MCMC sampling.
}
\section{Slots}{

\describe{
\item{\code{post}}{A named list containing a named list \code{par} with arrays for the
posterior density parameters.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputfixhier} for the parent class
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \linkS4class{mcmc} for the class defining the MCMC hyper-parameters
\item \code{\link[=mcmc]{mcmc()}} for the \code{mcmc} class constructor
}
}
