% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixpost.R
\docType{class}
\name{mcmcoutputfixpost-class}
\alias{mcmcoutputfixpost-class}
\alias{.mcmcoutputfixpost}
\title{Finmix \code{mcmcoutput} class for fixed indicators and posterior parameters}
\description{
The \code{mcmcoutputfixpost} class inherits from the \code{mcmcoutputfix} class and
adds a slot to store the parameters of the posterior distribution from which
the component parameters are drawn. The storage of posterior parameters is
controlled by the slot \code{storepost} in the \linkS4class{mcmc} class. If set
to \code{TRUE} posterior parameters are stored in the output of the MCMC sampling.
}
\section{Slots}{

\describe{
\item{\code{post}}{A named list containing a list \code{par} that contains the posterior
parameters as named arrays.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputfix} for the parent class
\item \linkS4class{mcmcoutputpost} for the corresponding class for unknown
indicators.
\item \linkS4class{mcmc} for the class defining the MCMC hyper-parameters
\item \code{\link[=mcmc]{mcmc()}} for the constructor of the \linkS4class{mcmc} class
}
}
\keyword{internal}
