% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhier.R
\docType{class}
\name{mcmcoutputhier-class}
\alias{mcmcoutputhier-class}
\alias{.mcmcoutputhier}
\title{Finmix \code{mcmcoutputhier} class}
\description{
This class inherits from the \code{mcmcoutputbase} class and stores draws from
MCMC sampling with unknown indicators and an hierarchical prior. It adds to
its parent class a slot for storing the parameters of the hierarchical prior.

To use an hierarchical prior in MCMC sampling the \code{prior} object needs to
have set slot \verb{@hier} to \code{TRUE}.
}
\section{Slots}{

\describe{
\item{\code{hyper}}{A named list containing the arrays with parameters from the
hierarchical prior.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputbase} for the parent class
\item \linkS4class{prior} for the class specifying the prior distribution
\item \code{\link[=prior]{prior()}} for the \code{prior} class constructor
\item \code{\link[=priordefine]{priordefine()}} for the advanced \code{prior} class constructor
}
}
