% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhierpost.R
\docType{class}
\name{mcmcoutputperm-class}
\alias{mcmcoutputperm-class}
\title{Finmix \code{mcmcoutputperm} class}
\description{
The mcmcoutputperm class stores MCMC samples after relabeling (permuting).
}
\details{
Calling \code{\link[=mcmcpermute]{mcmcpermute()}} on an \code{mcmcoutput} class permutes the labels of the
components
This class union includes all classes that define objects for permuted
MCMC samples and is used to dispatch methods for \code{mcmcoutputperm} objects.
}
