% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixhier.R,
%   R/mcmcoutputpermbase.R, R/mcmcoutputpermhier.R, R/mcmcoutputpermpost.R,
%   R/mcmcoutputpermhierpost.R
\docType{class}
\name{plotPostDens,mcmcoutputpermfixhier-method}
\alias{plotPostDens,mcmcoutputpermfixhier-method}
\alias{plotTraces,mcmcoutputpermbase-method}
\alias{plotHist,mcmcoutputpermbase-method}
\alias{plotDens,mcmcoutputpermbase-method}
\alias{plotPointProc,mcmcoutputpermbase-method}
\alias{plotSampRep,mcmcoutputpermbase-method}
\alias{plotPostDens,mcmcoutputpermbase-method}
\alias{plotTraces,mcmcoutputpermhier-method}
\alias{plotHist,mcmcoutputpermhier-method}
\alias{plotDens,mcmcoutputpermhier-method}
\alias{plotPointProc,mcmcoutputpermhier-method}
\alias{plotSampRep,mcmcoutputpermhier-method}
\alias{plotPostDens,mcmcoutputpermhier-method}
\alias{plotTraces,mcmcoutputpermpost-method}
\alias{plotHist,mcmcoutputpermpost-method}
\alias{plotDens,mcmcoutputpermpost-method}
\alias{plotPointProc,mcmcoutputpermpost-method}
\alias{plotSampRep,mcmcoutputpermpost-method}
\alias{plotPostDens,mcmcoutputpermpost-method}
\alias{mcmcoutputpermhierpost-class}
\alias{.mcmcoutputpermhierpost}
\alias{plotTraces,mcmcoutputpermhierpost-method}
\alias{plotHist,mcmcoutputpermhierpost-method}
\alias{plotDens,mcmcoutputpermhierpost-method}
\alias{plotPointProc,mcmcoutputpermhierpost-method}
\alias{plotSampRep,mcmcoutputpermhierpost-method}
\alias{plotPostDens,mcmcoutputpermhierpost-method}
\title{Plot posterior densities of the component parameters}
\usage{
\S4method{plotPostDens}{mcmcoutputpermfixhier}(x, dev = TRUE, ...)

\S4method{plotTraces}{mcmcoutputpermbase}(x, dev = TRUE, lik = 1, col = FALSE, ...)

\S4method{plotHist}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotDens}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotPointProc}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotSampRep}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotPostDens}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotTraces}{mcmcoutputpermhier}(x, dev = TRUE, lik = 1, col = FALSE, ...)

\S4method{plotHist}{mcmcoutputpermhier}(x, dev = TRUE, ...)

\S4method{plotDens}{mcmcoutputpermhier}(x, dev = TRUE, ...)

\S4method{plotPointProc}{mcmcoutputpermhier}(x, dev = TRUE, ...)

\S4method{plotSampRep}{mcmcoutputpermhier}(x, dev = TRUE, ...)

\S4method{plotPostDens}{mcmcoutputpermhier}(x, dev = TRUE, ...)

\S4method{plotTraces}{mcmcoutputpermpost}(x, dev = TRUE, lik = 1, col = FALSE, ...)

\S4method{plotHist}{mcmcoutputpermpost}(x, dev = TRUE, ...)

\S4method{plotDens}{mcmcoutputpermpost}(x, dev = TRUE, ...)

\S4method{plotPointProc}{mcmcoutputpermpost}(x, dev = TRUE, ...)

\S4method{plotSampRep}{mcmcoutputpermpost}(x, dev = TRUE, ...)

\S4method{plotPostDens}{mcmcoutputpermpost}(x, dev = TRUE, ...)

\S4method{plotTraces}{mcmcoutputpermhierpost}(x, dev = TRUE, lik = 1, col = FALSE, ...)

\S4method{plotHist}{mcmcoutputpermhierpost}(x, dev = TRUE, ...)

\S4method{plotDens}{mcmcoutputpermhierpost}(x, dev = TRUE, ...)

\S4method{plotPointProc}{mcmcoutputpermhierpost}(x, dev = TRUE, ...)

\S4method{plotSampRep}{mcmcoutputpermhierpost}(x, dev = TRUE, ...)

\S4method{plotPostDens}{mcmcoutputpermhierpost}(x, dev = TRUE, ...)
}
\arguments{
\item{x}{An \code{mcmcoutputpermhierpost} object containing all sampled values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{...}{Further arguments to be passed to the plotting function.}

\item{lik}{An integer indicating, if the log-likelihood traces should be
plotted (default). If set to \code{0} the traces for the parameters
and weights are plotted instead.}

\item{col}{A logical indicating, if the plot should be colored.}
}
\value{
Posterior densities of the MCMC samples.

A plot of the traces of the MCMC samples.

Histograms of the MCMC samples.

Densities of the MCMC samples.

Point process of the MCMC samples.

Sampling representation of the MCMC samples.

Posterior densities of the MCMC samples.

A plot of the traces of the MCMC samples.

Histograms of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Posterior densities of the MCMC samples.

A plot of the traces of the MCMC samples.

Histograms of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Posterior densities of the MCMC samples.

A plot of the traces of the MCMC samples.

Histograms of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Densities of the MCMC samples.

Posterior densities of the MCMC samples.
}
\description{
Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson and Binomial
mixture distributions.

Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{1}.

If \code{lik} is set to \code{0} the parameters of the components and the posterior
parameters are plotted together with \code{K-1} weights.

Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling.

Note, this method is so far only implemented for mictures of Poisson and
Binomial distributions.

Calling \code{\link[=plotDens]{plotDens()}} plots densities of the sampled parameters and weights
from MCMC sampling.

Calling \code{\link[=plotPointProc]{plotPointProc()}} plots point processes of the sampled component
parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotSampRep]{plotSampRep()}} plots sampling representations of the sampled
component parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson or Binomial
mixture distributions.

Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{1}.

If \code{lik} is set to \code{0} the parameters of the components and the hierarchical
prior are plotted together with \code{K-1} weights.

Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling. In addition the parameters of the hierarchical prior are
plotted.

Note, this method is so far only implemented for mictures of Poisson and
Binomial distributions.

Calling \code{\link[=plotDens]{plotDens()}} plots densities of the sampled parameters and weights
from MCMC sampling. In addition, the parameters of the hierarchical prior
are plotted.

Note, this method is so far only implemented for mixtures of Poisson and
Binomial distributions.

Calling \code{\link[=plotPointProc]{plotPointProc()}} plots point processes of the sampled component
parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotSampRep]{plotSampRep()}} plots sampling representations of the sampled
component parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson or Binomial
mixture distributions.

Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{1}.

If \code{lik} is set to \code{0} the parameters of the components and the posterior
parameters are plotted together with \code{K-1} weights.

Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling.

Note, this method is so far only implemented for mixtures of Poisson and
Binomial distributions.

Calling \code{\link[=plotDens]{plotDens()}} plots densities of the sampled parameters and weights
from MCMC sampling.

Note, this method is so far only implemented for mixtures of Poisson and
Binomial distributions.

Calling \code{\link[=plotPointProc]{plotPointProc()}} plots point processes of the sampled component
parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotSampRep]{plotSampRep()}} plots sampling representations of the sampled
component parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson or Binomial
mixture distributions.

This class defines objects to store the outputs from permuting the MCMC
samples. Due to label switching the sampled component parameters are usually
not assigned to the same component in each iteration. To overcome this issue
the samples are permuted by using a relabeling algorithm (usually K-means)
to reassign parameters. Note that due to assignment of parameters from the
same iteration to the same component, the sample size could shrink.

This class stores the permuted parameters together with the new MCMC sample
size and the mixture log-likelihood, the prior log-likelihood, and the
complete data posterior log-likelihood.

Note that this class inherits all of its slots from the parent classes.

Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{1}.

If \code{lik} is set to \code{0} the parameters of the components, the posterior
parameters, and the parameters of the hierarchical prior are plotted
together with \code{K-1} weights.

Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling. In addition the parameters of the hierarchical prior are
plotted.

Note, this method is so far only implemented for mictures of Poisson and
Binomial distributions.

Calling \code{\link[=plotDens]{plotDens()}} plots densities of the sampled parameters and weights
from MCMC sampling. In addition, the parameters of the hierarchical prior
are plotted.

Note, this method is so far only implemented for mixtures of Poisson and
Binomial distributions.

Calling \code{\link[=plotPointProc]{plotPointProc()}} plots point processes of the sampled component
parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotSampRep]{plotSampRep()}} plots sampling representations of the sampled
component parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson or Binomial
mixture distributions.
}
\section{Functions}{
\itemize{
\item \code{plotPostDens,mcmcoutputpermfixhier-method}: 

\item \code{plotTraces,mcmcoutputpermbase-method}: 

\item \code{plotHist,mcmcoutputpermbase-method}: 

\item \code{plotDens,mcmcoutputpermbase-method}: 

\item \code{plotPointProc,mcmcoutputpermbase-method}: 

\item \code{plotSampRep,mcmcoutputpermbase-method}: 

\item \code{plotPostDens,mcmcoutputpermbase-method}: 

\item \code{plotTraces,mcmcoutputpermhier-method}: Plots traces of MCMC samples

\item \code{plotHist,mcmcoutputpermhier-method}: 

\item \code{plotDens,mcmcoutputpermhier-method}: 

\item \code{plotPointProc,mcmcoutputpermhier-method}: 

\item \code{plotSampRep,mcmcoutputpermhier-method}: 

\item \code{plotPostDens,mcmcoutputpermhier-method}: 

\item \code{plotTraces,mcmcoutputpermpost-method}: 

\item \code{plotHist,mcmcoutputpermpost-method}: Plots histograms of MCMC samples

\item \code{plotDens,mcmcoutputpermpost-method}: Plots densities of MCMC samples

\item \code{plotPointProc,mcmcoutputpermpost-method}: Plots point process of MCMC samples

\item \code{plotSampRep,mcmcoutputpermpost-method}: Plots sampling representations of MCMC
samples

\item \code{plotPostDens,mcmcoutputpermpost-method}: Plots posterior densities of component
parameters

\item \code{mcmcoutputpermhierpost-class}: Finmix \code{mcmcoutputpermhierpost} class

\item \code{plotTraces,mcmcoutputpermhierpost-method}: 

\item \code{plotHist,mcmcoutputpermhierpost-method}: 

\item \code{plotDens,mcmcoutputpermhierpost-method}: 

\item \code{plotPointProc,mcmcoutputpermhierpost-method}: 

\item \code{plotSampRep,mcmcoutputpermhierpost-method}: 

\item \code{plotPostDens,mcmcoutputpermhierpost-method}: 
}}

\examples{
\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2, 
                 indicfix = TRUE)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)
}

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotTraces(f_outputperm, lik = 0)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPointProc(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotSampRep(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotTraces(f_outputperm, lik = 0)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPointProc(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotSampRep(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotTraces(f_outputperm, lik = 0)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPointProc(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotSampRep(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotTraces(f_outputperm, lik = 0)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotDens(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPointProc(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotSampRep(f_outputperm)

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)

}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}

\itemize{
\item \link[=mcmcoutput_class]{mcmcoutputbase} for the parent class
\item \link[=mcmcperm_class]{mcmcpermind} for the parent class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for performing permutation of MCMC samples
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}
}
