% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermbase.R
\docType{class}
\name{mcmcoutputpermbase-class}
\alias{mcmcoutputpermbase-class}
\alias{.mcmcoutputpermbase}
\alias{show,mcmcoutputpermbase-method}
\alias{plotTraces,mcmcoutputpermbase-method}
\alias{plotHist,mcmcoutputpermbase-method}
\alias{plotDens,mcmcoutputpermbase-method}
\alias{plotPointProc,mcmcoutputpermbase-method}
\alias{plotSampRep,mcmcoutputpermbase-method}
\alias{plotPostDens,mcmcoutputpermbase-method}
\title{Finmix \code{mcmcoutputpermbase} class}
\usage{
\S4method{show}{mcmcoutputpermbase}(object)

\S4method{plotTraces}{mcmcoutputpermbase}(x, dev = TRUE, lik = 1, col = FALSE, ...)

\S4method{plotHist}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotDens}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotPointProc}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotSampRep}{mcmcoutputpermbase}(x, dev = TRUE, ...)

\S4method{plotPostDens}{mcmcoutputpermbase}(x, dev = TRUE, ...)
}
\arguments{
\item{object}{An \code{mcmcoutputpermbase} object.}

\item{x}{An \code{mcmcoutputpermbase} object containing all sampled values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{lik}{An integer indicating, if the log-likelihood traces should be
plotted (default). If set to \code{0} the traces for the parameters
and weights are plotted instead.}

\item{col}{A logical indicating, if the plot should be colored.}

\item{...}{Further arguments to be passed to the plotting function.}
}
\value{
A console output listing the slots and summary information about
each of them.

A plot of the traces of the MCMC samples.

Histograms of the MCMC samples.

Densities of the MCMC samples.

Point process of the MCMC samples.

Sampling representation of the MCMC samples.

Posterior densities of the MCMC samples.
}
\description{
This class defines objects to store the outputs from permuting the MCMC
samples. Due to label switching the sampled component parameters are usually
not assigned to the same component in each iteration. To overcome this issue
the samples are permuted by using a relabeling algorithm (usually K-means)
to reassign parameters. Note that due to assignment of parameters from the
same iteration to the same component, the sample size could shrink.

This class stores the permuted parameters together with the new MCMC sample
size and the mixture log-likelihood, the prior log-likelihood, and the
complete data posterior log-likelihood.

Note that this class inherits all of its slots from the parent classes.

Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermbase} object gives an overview
of the \code{mcmcoutputpermbase} object.

Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{1}.

If \code{lik} is set to \code{0} the parameters of the components and the posterior
parameters are plotted together with \code{K-1} weights.

Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling.

Note, this method is so far only implemented for mictures of Poisson and
Binomial distributions.

Calling \code{\link[=plotDens]{plotDens()}} plots densities of the sampled parameters and weights
from MCMC sampling.

Calling \code{\link[=plotPointProc]{plotPointProc()}} plots point processes of the sampled component
parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotSampRep]{plotSampRep()}} plots sampling representations of the sampled
component parameters from MCMC sampling.

Note, this method is only implemented for mixtures of Poisson and Binomial
distributions.

Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this method is so far only implemented for Poisson or Binomial
mixture distributions.
}
\section{Functions}{
\itemize{
\item \code{show,mcmcoutputpermbase-method}: Shows a short summary of the object's
slots

\item \code{plotTraces,mcmcoutputpermbase-method}: 

\item \code{plotHist,mcmcoutputpermbase-method}: Plot histograms of the parameters and
weights

\item \code{plotDens,mcmcoutputpermbase-method}: 

\item \code{plotPointProc,mcmcoutputpermbase-method}: Plots point process for the component
parameters

\item \code{plotSampRep,mcmcoutputpermbase-method}: Plots sampling representations of the
component parameters

\item \code{plotPostDens,mcmcoutputpermbase-method}: Plots the posterior density of the
component parameters
}}

\examples{
\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotTraces(f_outputperm, lik = 0)
}

\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)
}

# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotDens(f_outputperm)

\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPointProc(f_outputperm)
}

\dontrun{ 
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotSampRep(f_outputperm)
}

\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Do not use a hierarchical prior.
setHier(f_prior) <- FALSE
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotPostDens(f_outputperm)
}

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputbase} for the parent class
\item \linkS4class{mcmcpermind} for the parent class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for performing permutation of MCMC samples
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes of sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting posterior densities for sampled values
}

\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}
}
