% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpost.R
\docType{class}
\name{mcmcoutputpost-class}
\alias{mcmcoutputpost-class}
\alias{.mcmcoutputpost}
\title{Finmix \code{mcmcoutputpost} class}
\description{
This class inherits from the \code{mcmcoutputbase} class and adds posterior
density parameters to the MCMC sampling output. The storage of posterior
parameters is controlled by the slot \code{storepost} in the \link[=mcmc_class]{mcmc}
class. If set to \code{TRUE} posterior parameters are stored in the output of the
MCMC sampling.
}
\section{Slots}{

\describe{
\item{\code{post}}{A named list containing a named list \code{par} with arrays for the
posterior density parameters.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputbase} for the parent class
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling for finite mixture modeling
}
}
