% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfix.R
\docType{class}
\name{mcmcpermfix-class}
\alias{mcmcpermfix-class}
\alias{.mcmcpermfix}
\title{Finmix \code{mcmcpermfix} class}
\description{
This class defines objects to store the outputs from permuting the MCMC
samples. Due to label switching the sampled component parameters are usually
not assigned to the same component in each iteration. To overcome this issue
the samples are permuted by using a relabeling algorithm (usually K-means)
to reassign parameters. Note that due to assignment of parameters from the
same iteration to the same component, the sample size could shrink.

This class stores the permuted parameters together with the new MCMC sample
size and the mixture log-likelihood, the prior log-likelihood, and the
complete data posterior log-likelihood.

Note that for models with fixed indicators \code{weight}s do not get permuted.
}
\section{Slots}{

\describe{
\item{\code{Mperm}}{An integer storing the MCMC sample size after relabeling.}

\item{\code{parperm}}{A named list containing the permuted component parameters.}

\item{\code{logperm}}{A named list containing the mixture log-likelihood, the prior
log-likelihood, and the complete data posterior log-likelihood.}
}}

\seealso{
\itemize{
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for the calling function
\item \linkS4class{mcmcpermind} for the corresponding class for models with
unknown indicators
}
}
\concept{mcmcperm-classes}
