% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfixhier.R
\docType{methods}
\name{getHyperperm,mcmcpermfixpost-method}
\alias{getHyperperm,mcmcpermfixpost-method}
\alias{mcmcpermfixhierpost_class,}
\alias{mcmcoutputpermfixhier_class,}
\alias{mcmcpermoutputpermfixhierpost_class}
\title{Getter method of \code{mcmcpermfixhier} class.}
\usage{
\S4method{getHyperperm}{mcmcpermfixpost}(object)
}
\arguments{
\item{object}{An \code{mcmcpermfixhier} object.}
}
\value{
The \code{hyperperm} slot of the \code{object}.
}
\description{
Returns the \code{hyperperm} slot.
}
\examples{
\dontrun{getHyperpem(mcmcperm)}

}
\seealso{
\itemize{
\item \code{\link{mcmcoutputpermfix-class}} for the inheriting class
\item \code{\link{mcmcpermute}} for function permuting MCMC samples
}
}
