% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcpermfixpost.R
\docType{class}
\name{mcmcpermfixpost-class}
\alias{mcmcpermfixpost-class}
\alias{.mcmcpermfixpost}
\title{Finmix \code{mcmcpermfixpost} class}
\description{
This class defines objects to store the outputs from permuting the MCMC
samples. Due to label switching the sampled component parameters are usually
not assigned to the same component in each iteration. To overcome this issue
the samples are permuted by using a relabeling algorithm (usually K-means)
to reassign parameters. Note that due to assignment of parameters from the
same iteration to the same component, the sample size could shrink.

This class is supplementing the parent class by adding a slot to store the
permuted parameter samples of the posterior densities.

Note that for models with fixed indicators \code{weight}s do not get permuted.
}
\section{Slots}{

\describe{
\item{\code{postperm}}{A named list containing a named list \code{par} with array(s) of
parameters from the posterior density.}
}}

\seealso{
\itemize{
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for the calling function
\item \linkS4class{mcmcpermfix} for the parent class definition
\item \linkS4class{mcmcpermindpost}for the corresponding class for models with
unknown indicators
}
}
