% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcstart.R
\name{mcmcstart}
\alias{mcmcstart}
\title{Finmix default starting values}
\usage{
mcmcstart(fdata, model, varargin)
}
\arguments{
\item{fdata}{An \code{fdata} object containing the data.}

\item{model}{A \code{model} object specifying the finite mixture model to be
estimated.}

\item{varargin}{Either \code{NULL} or an \code{mcmc} object defining (possibly
non-default) hyper-parameters. If not provided a default \code{mcmc} object is
created internally and returned.}
}
\value{
A list containing the \code{fdata} object, the \code{model} object and an
\code{mcmc} object all set up for MCMC sampling.
}
\description{
Calling \code{\link[=mcmcstart]{mcmcstart()}} creates starting values for MCMC sampling. Starting
values are constructed for the indicators in the \code{fdata} argument and the
parameters in the \code{model} argument. In addition an \code{mcmc} object can be
provided in the \code{varargin} argument to set up all slots consistently for a
non-default setting of hyper-parameters.

To assing the returned objects directly to existing names the assignment
operator \verb{\%\%=\%\%} can be used together with a formula concatenating each name
with a tilde \code{~}. See the examples.
}
\examples{
# Specify a Poisson model.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Set up all objects for MCMC sampling.
(f_data ~ f_model ~ f_mcmc) \%=\% mcmcstart(f_data, f_model)

}
\seealso{
\itemize{
\item \link[=fdata_class]{fdata} for the definition of the \code{fdata} class
\item \link[=model_class]{model} for the definition of the \code{model} class
\item \link[=mcmc_class]{mcmc} for the definition of the \code{mcmc} class
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for the starting MCMC sampling
}
}
