% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{mixturemar,model-method}
\alias{mixturemar,model-method}
\title{Returns the marginal distribution.}
\usage{
\S4method{mixturemar}{model}(object, J)
}
\arguments{
\item{object}{An S4 model object with a multivariate distribution.}

\item{J}{An integer specifying the dimension for which the marginal
distribution should be returned.}
}
\value{
An S4 model object with the marginal distribution for dimension
\code{J}.
}
\description{
\code{mixturemar} returns the marginal distribution of a multivariate
mixture distribution. This can only be applied on S4 model objects with
\code{dist="normult"} or \code{dist="studmult"}.
}
\examples{
\dontrun{
mar_model <- mixturemar(f_model, 1)
}

}
\seealso{
\code{model}
}
