% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{mixturemar}
\alias{mixturemar}
\title{Extracts the marginal distribution from a finite mixture model}
\usage{
mixturemar(object, J)
}
\description{
Extracts the marginal distribution from a finite mixture model
}
\keyword{internal}
