% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{Constructor for the S4 model class}
\usage{
model(
  dist = "poisson",
  r,
  K,
  weight = matrix(),
  par = list(),
  indicmod = "multinomial",
  indicfix = FALSE,
  T = matrix()
)
}
\arguments{
\item{dist}{A character, defining the distribution family. Possible choices
are \code{"binomial"}, \code{"exponential"}, \code{"normal"},
\code{"normult"}, \code{"poisson"}, \code{"student"}, and \code{"studmult"}.}

\item{r}{An integer. Defines the vector dimension of a model. Is one for all
univariate distributions and larger than one for \code{"normult"} and
\code{"studmult"}.}

\item{K}{An integer, defining the number of components in the finite mixture.
Must be larger or equal to one.}

\item{weight}{A matrix, containing the weights of the finite mixture model.
The matrix must have dimension \code{1 x K} and weights must add to one
and must all be larger or equal to zero.}

\item{par}{A list containing the parameter vectors for the finite mixture
distribution. The list can contain more than one named parameter vector.
Depending on the distribution parameters must be defined in the list as
follows: a \code{K}-dimensional vector of probabilities named \code{"p"} for
a \code{"binomial"} model, a \code{K}-dimensional vector of positive rates
named \code{"lambda"} for an \code{"exponential"} model,
\code{K}-dimensional vectors of means named \code{"mu"} and variances named
\code{sigma} for a \code{"normal"} model, a \code{r x K}-dimensional
matrix of means named \code{"mu"} and a \code{K x r x r} dimensional
array of variance-covariance matrices named \code{"sigma"} for a
\code{"normult"} model, a \code{K}-dimensional vector of rates named
\code{"rates"} for a \code{"poisson"} model, \code{K}-dimensional vectors of
means named \code{"mu"}, variances named \code{sigma}, and degrees of freedom
named \code{"df"} for a \code{"student"} model, a
\code{r x K}-dimensional matrix of means named \code{"mu"}, a
\code{K x r x r} dimensional array of variance-covariance matrices
named \code{"sigma"}, and a \code{K}-dimensional vector of degrees of freedom
for a \code{"studmult"} model.}

\item{indicmod}{A character defining the indicator model used. For now only
\code{"multinomial"} is implemented.}

\item{indicfix}{A logical. If \code{TRUE} the indicators are given and
therefore fixed.}

\item{T}{A matrix containing the repetitions in case of a \code{"binomial"} or
\code{"poisson"} model. Must be positive integers.}
}
\value{
An S4 \code{model} object.
}
\description{
\code{model} creates a finite mixture model with given parameters.
}
\details{
This is a constructor that creates a class object and guides the user in
regard to the different parameters needed to define a finite mixture model.
}
\examples{
f_model <- model(dist = "poisson", K = 2, par = list(lambda = c(0.17, 0.2)))

}
\seealso{
\itemize{
\item \link[=model-class]{model} for the class definition
}
}
