% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasWeight,model-method}
\alias{hasWeight,model-method}
\alias{model_class}
\title{Getter for weights}
\usage{
\S4method{hasWeight}{model}(object, verbose = FALSE)

\S4method{getDist}{model}(object)

\S4method{getR}{model}(object)

\S4method{getK}{model}(object)

\S4method{getWeight}{model}(object)

\S4method{getPar}{model}(object)

\S4method{getIndicmod}{model}(object)

\S4method{getIndicfix}{model}(object)

\S4method{getT}{model}(object)

\S4method{setDist}{model}(object) <- value

\S4method{setR}{model}(object) <- value

\S4method{setK}{model}(object) <- value

\S4method{setWeight}{model}(object) <- value

\S4method{setPar}{model}(object) <- value

\S4method{setIndicmod}{model}(object) <- value

\S4method{setIndicfix}{model}(object) <- value

\S4method{setT}{model}(object) <- value
}
\arguments{
\item{verbose}{A logical indicating, if the function should give a print out.}

\item{model}{An S4 model object.}
}
\value{
Matrix of weights.
}
\description{
\code{hasWeight} returns the weight matrix.
}
\examples{
\dontrun{
weight <- hasWeight(model)
}
}
