% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmoments.R
\docType{class}
\name{modelmoments-class}
\alias{modelmoments-class}
\title{Finmix \code{modelmoments} class}
\description{
Defines a container to hold the moments of a finite mixture model. The
finmix \code{model} object should contains parameters and weights.
}
\section{Slots}{

\describe{
\item{\code{mean}}{A vector of component means.}

\item{\code{var}}{An array of components variances or in case of multivariate
distributions covariance matrices.}

\item{\code{model}}{The corresponding \code{model} object.}
}}

\seealso{
\itemize{
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of the \code{modelmoments} class
}
}
