% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmoments.R
\name{modelmoments}
\alias{modelmoments}
\title{Constructor of finmix \code{modelmoments} class}
\usage{
modelmoments(model)
}
\arguments{
\item{model}{A \code{model} object containing defined parameters in slot \code{par}
and defined weights in slot \code{weight}.}
}
\value{
A \code{modelmoments} object with calculated moments of the finite
mixture model defined in the \code{model} object.
}
\description{
Calling \code{\link[=modelmoments]{modelmoments()}} calculates the corresponding moments of the
finite mixture distribution defined in the \code{model} object. The \code{model}
object should contain parameters in slot \code{par} and weights in slot \code{weight}.
}
\examples{
f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
modelmoments(f_model)

}
\seealso{
\itemize{
\item \link{modelmoments_class} for all slots of the \code{modelmoments} class
}
}
