% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelmoments.R
\docType{class}
\name{modelmoments_class}
\alias{modelmoments_class}
\alias{getMean,modelmoments-method}
\alias{getVar,modelmoments-method}
\alias{getModel,modelmoments-method}
\title{Finmix \code{modelmoments} class}
\usage{
\S4method{getMean}{modelmoments}(object)

\S4method{getVar}{modelmoments}(object)

\S4method{getModel}{modelmoments}(object)
}
\arguments{
\item{object}{A \code{modelmoments} object.}
}
\value{
The \code{mean} slot of the \code{object}.

The \code{var} slot of the \code{object}.

The \code{model} slot of the \code{object}.
}
\description{
Defines a container to hold the moments of a finite mixture model. The
finmix \code{model} object should contains parameters and weights.

Returns the \code{mean} slot of a \code{modelmoments} object.

Returns the \code{var} slot of a \code{modelmoments} object.

Returns the \code{model} slot of a \code{modelmoments} object.
}
\section{Functions}{
\itemize{
\item \code{getMean,modelmoments-method}: 

\item \code{getVar,modelmoments-method}: 

\item \code{getModel,modelmoments-method}: 
}}

\section{Slots}{

\describe{
\item{\code{mean}}{A vector of component means.}

\item{\code{var}}{An array of components variances or in case of multivariate
distributions covariance matrices.}

\item{\code{model}}{The corresponding \code{model} object.}
}}

\examples{
f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getMean(f_moments)

f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getVar(f_moments)

f_model <- model("poisson", par=list(lambda=c(0.3, 0.1)), 
                 weight=matrix(c(0.3, 0.7), nrow=1))
f_moments <- modelmoments(f_model)
getModel(f_moments)

}
\seealso{
\itemize{
\item \code{\link[=modelmoments]{modelmoments()}} the constructor of the \code{modelmoments} class
}

\link{modelmoments_class} for all slots of the \code{modelmoments} class

\link{modelmoments_class} for all slots of the \code{modelmoments} class

\link{modelmoments_class} for all slots of the \code{modelmoments} class
}
