% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcextract.R
\name{moments,mcmcextract-method}
\alias{moments,mcmcextract-method}
\title{Calculate the model moments of MCMC samples}
\usage{
\S4method{moments}{mcmcextract}(object)
}
\arguments{
\item{obj}{An \code{mcmcextract} object containing the parameters and weights
from MCMC sampling.}
}
\value{
A list containing the model moments calculated from MCMC samples.
}
\description{
For internal usage only. This function calculates the finite mixture moments
of a mixture model from the MCMC samples. Note that this function is
momentarily only implemented for a mixture of multivariate Normal
distributions.
}
\seealso{
\itemize{
\item \linkS4class{mcmcoutput} for the results from MCMC sampling
\item \code{\link[=extract]{extract()}} for the calling method
}
}
\keyword{internal}
