% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfix.R
\name{moments,mcmcoutputfix-method}
\alias{moments,mcmcoutputfix-method}
\title{Computes multivariate Normal sample moments}
\usage{
\S4method{moments}{mcmcoutputfix}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputfix} object containing all data from MCMC
sampling.}
}
\value{
The moments on the samples of a multivariate Normal mixture.
}
\description{
Calling \code{moments()} calculates the sample moments for the samples of a
multivariate Normal mixture model.
}
\keyword{internal}
