% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{moments}
\alias{moments}
\title{Computes the model moments from MCMC samples}
\usage{
moments(object)
}
\description{
Computes the model moments from MCMC samples
}
\keyword{internal}
