% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moments_cc}
\alias{moments_cc}
\title{Calculate moments on samples of multivariate mixture models}
\usage{
moments_cc(classS4)
}
\arguments{
\item{classS4}{An \code{mcmcoutput} class containing the MCMC samples.}
}
\value{
A named list with vectors containing the data moments for each
iteration in the MCMC sample.
}
\description{
This function calculates the moments for MCMC samples of multivariate
mixture models. Moments like means, standard deviations, kurtosis and
skewness are computed for each iteration in MCMC sampling. The moments are
used when plotting the traces of an MCMC sample output.
}
\seealso{
\itemize{
\item \linkS4class{mcmcoutput} for the \code{mcmcoutput} class definition
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=plotTraces]{plotTraces()}} for the calling function
}
}
