% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmodelmoments.R
\docType{class}
\name{normalmodelmoments}
\alias{normalmodelmoments}
\alias{.normalmodelmoments}
\title{Finmix \code{normalmodelmoments} class}
\description{
Defines a class that holds theoretical moments for a finite mixture of
normal distributions. Note that this class is not directly used, but
indirectly when calling the \code{modelmoments} constructor \code{\link[=modelmoments]{modelmoments()}}.
}
\section{Slots}{

\describe{
\item{\code{B}}{A numeric defining the between-group heterogeneity.}

\item{\code{W}}{A numeric defining the within-group heterogeneity.}

\item{\code{R}}{A numeric defining the coefficient of determination.}
}}

\seealso{
\itemize{
\item \link{modelmoments_class} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
}
}
