% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{permmoments_cc}
\alias{permmoments_cc}
\title{Calculate moments on permuted samples of multivariate mixture models}
\usage{
permmoments_cc(classS4)
}
\arguments{
\item{classS4}{An \code{mcmcoutputperm} class containing the re-labeled MCMC
samples.}
}
\value{
A named list with vectors containing the data moments for each
iteration in the re-labeled MCMC sample.
}
\description{
This function calculates the moments for re-labeled MCMC samples of
multivariate mixture models. Moments like means, standard deviations,
kurtosis and skewness are computed for each iteration in MCMC sampling. The
moments are used when plotting the traces of an MCMC sample output.
}
\seealso{
\itemize{
\item \link[=mcmcoutputperm_class]{mcmcoutputperm} for the \code{mcmcoutput} class definition
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for re-labeling MCMC samples
\item \link[=mcmcoutputperm_class]{plotTraces} for the calling function
}
}
