% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{plot,model,ANY-method}
\alias{plot,model,ANY-method}
\title{Plots a model.}
\usage{
\S4method{plot}{model,ANY}(x, y, dev = TRUE, ...)
}
\arguments{
\item{x}{An S4 model object. Must have specified parameters and weights.}

\item{y}{Unused.}

\item{dev}{A logical indicating, if the plot should be shown in a graphical
device. Set to \code{FALSE}, if plotted to a file.}
}
\value{
Density or barplot of the S4 model object.
}
\description{
\code{plot} plots the density or probabilities of a fully specified mixture
model.
}
\examples{
\dontrun{
plot(f_model)
}

}
\seealso{
\code{model}
}
