% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotHist-generic}
\alias{plotHist-generic}
\title{Plots histograms of MCMC samples}
\usage{
plotHist(x, dev = TRUE, ...)
}
\description{
Plots histograms of MCMC samples
}
\keyword{internal}
