% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixhierpost.R
\name{plotHist,mcmcoutputpermfixhierpost-method}
\alias{plotHist,mcmcoutputpermfixhierpost-method}
\title{Plot histograms of the parameters and weights}
\usage{
\S4method{plotHist}{mcmcoutputpermfixhierpost}(x, dev = TRUE, ...)
}
\arguments{
\item{x}{An \code{mcmcoutputpermfixhierpost} object containing all sampled values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{...}{Further arguments to be passed to the plotting function.}
}
\value{
Histograms of the MCMC samples.
}
\description{
Calling \code{\link[=plotHist]{plotHist()}} plots histograms of the sampled parameters and weights
from MCMC sampling.More specifically, all component parameters, \code{K-1} of the
weights and the posterior parameters are considered in the histogram plots.

Note, this method is so far only implemented for Poisson and Binomial
mixture distributions.
}
\examples{
\dontrun{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2, 
                 indicfix = TRUE)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
f_outputperm <- mcmcpermute(f_output)
plotHist(f_outputperm)
}

}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}
}
