% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotPointProc-generic}
\alias{plotPointProc-generic}
\title{Plots the point process of a finite mixture model}
\usage{
plotPointProc(x, dev = TRUE, ...)
}
\description{
Plots the point process of a finite mixture model
}
\keyword{internal}
