% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotPostDens-generic}
\alias{plotPostDens-generic}
\title{Plots the posterior density of sampled component parameters}
\usage{
plotPostDens(x, dev = TRUE, ...)
}
\description{
Plots the posterior density of sampled component parameters
}
\keyword{internal}
