% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixpost.R
\name{plotPostDens,mcmcoutputfixpost-method}
\alias{plotPostDens,mcmcoutputfixpost-method}
\title{Plot posterior densities of the component parameters}
\usage{
\S4method{plotPostDens}{mcmcoutputfixpost}(x, dev = TRUE, ...)
}
\arguments{
\item{x}{An \code{mcmcoutput} object containing all sampled values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{...}{Further arguments to be passed to the plotting function.}
}
\value{
Posterior densities of the MCMC samples.
}
\description{
Calling \code{\link[=plotPostDens]{plotPostDens()}} plots posterior densities of the sampled component
parameters from MCMC sampling, if the number of components is two.

Note, this methid calls the equivalent method of the parent class
\code{mcmcoutputfix}.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2, 
                 indicfix = TRUE)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
plotPostDens(f_output)

}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
}
}
\keyword{internal}
