% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhierpost.R
\name{plotPostDens}
\alias{plotPostDens}
\title{Plot the posterior density of component parameters}
\arguments{
\item{x}{An \code{mcmcoutput} or \code{mcmcoutputperm} object containing all sampled
values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{...}{Further arguments to be passed to the plotting function.}
}
\value{
The posterior density of the MCMC samples.
}
\description{
Calling \code{\link[=plotPostDens]{plotPostDens()}} on an object of class \code{mcmcoutput} or
\code{mcmcoutputperm} plots the posterior density of the sampled component
parameters from MCMC sampling, either the original parameters or the
relabeled ones (\code{mcmcoutputperm}).
}
\details{
Next to sampling representations and the point process of MCMC samples the
posterior density of component parameters can also be plotted directly for
finite mixture distributions with \code{ K=2} components and a single parameter.
The posterior density will always be bimodal due to to label-switching in
the MCMC sampling. This could change when considering a relabeld MCMC sample
(\code{mcmcoutputperm} object).

Note that this method for \code{mcmcoutputperm} objects is only implemented for
mixtures of Poisson and Binomial distributions.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc()
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
plotPostDens(f_output)

}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for permuting MCMC samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting the traces of sampled values
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting the point process of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting the sampling representation for sampled values
}
}
