% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotSampRep}
\alias{plotSampRep}
\title{Plots sample representations of MCMC samples}
\usage{
plotSampRep(x, dev = TRUE, ...)
}
\description{
Plots sample representations of MCMC samples
}
\keyword{internal}
