% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{plotTraces}
\alias{plotTraces}
\title{Plots the traces of the MCMC samples}
\usage{
plotTraces(x, dev = TRUE, lik = 1, col = FALSE, ...)
}
\description{
Plots the traces of the MCMC samples
}
\keyword{internal}
