% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhier.R
\name{plotTraces,mcmcoutputfixhier-method}
\alias{plotTraces,mcmcoutputfixhier-method}
\title{Plot traces of MCMC sampling}
\usage{
\S4method{plotTraces}{mcmcoutputfixhier}(x, dev = TRUE, lik = 1, col = FALSE, ...)
}
\arguments{
\item{x}{An \code{mcmcoutput} object containing all sampled values.}

\item{dev}{A logical indicating, if the plots should be shown by a graphical
device. If plots should be stored to a file set \code{dev} to \code{FALSE}.}

\item{lik}{An integer indicating, if the log-likelihood traces should be
plotted (default). If set to \code{0} the traces for the parameters
and weights are plotted instead.}

\item{col}{A logical indicating, if the plot should be colored.}

\item{...}{Further arguments to be passed to the plotting function.}
}
\value{
A plot of the traces of the MCMC samples.
}
\description{
Calling \code{\link[=plotTraces]{plotTraces()}} plots the MCMC traces of the mixture log-likelihood
, the mixture log-likelihood of the prior distribution, the log-likelihood
of the complete data posterior, or the weights and parameters if \code{lik} is
set to \code{0}.

If \code{lik} is set to \code{0} the parameters of the components and the posterior
parameters are plotted together with \code{K-1} weights.
}
\examples{
# Define a Poisson mixture model with two components.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2, 
                 indicfix = TRUE)
# Simulate data from the mixture model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
plotTraces(f_output, lik = 0)

}
\seealso{
\itemize{
\item \code{\link[=mixturemcmc]{mixturemcmc()}} for performing MCMC sampling
\item \code{\link[=plotHist]{plotHist()}} for plotting histograms of sampled values
\item \code{\link[=plotDens]{plotDens()}} for plotting densities of sampled values
\item \code{\link[=plotSampRep]{plotSampRep()}} for plotting sampling representations of sampled values
\item \code{\link[=plotPointProc]{plotPointProc()}} for plotting point processes for sampled values
\item \code{\link[=plotPostDens]{plotPostDens()}} for plotting the posterior density of component parameters
}
}
\keyword{internal}
