% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonmodelmoments.R
\docType{class}
\name{poissonmodelmoments-class}
\alias{poissonmodelmoments-class}
\alias{.poissonmodelmoments}
\title{Finmix \code{poissonmodelmoments} class}
\description{
Defines a class that holds modelmoments for a finite mixture of poisson
distributions. Note that this class is not directly used, but indirectly
when calling the \code{modelmoments} constructor \code{\link[=modelmoments]{modelmoments()}}.
}
\seealso{
\itemize{
\item \linkS4class{dmodelmoments} for the parent class
\item \linkS4class{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
}
}
\keyword{internal}
