% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{prior}
\alias{prior}
\title{Constructor for \code{prior} class}
\usage{
prior(
  weight = matrix(),
  par = list(),
  type = c("independent", "condconjugate"),
  hier = TRUE
)
}
\description{
Calling \code{\link[=prior]{prior()}} constructs an object of class \link[=prior-class]{prior}. The
constructor can be called without providing any arguments, but the prior
has to be filled with appropriate parameters when MCMC sampling should be
performed.

There exists next to the general constructor also an advanced constructor
that specifies a data dependent prior. See \code{\link[=priordefine]{priordefine()}} for this advanced
constructor.
}
\section{Slots}{

\describe{
\item{\code{weight}}{A matrix storing the prior parameters for the \code{weight} of a
finite mixture model.}

\item{\code{par}}{A list storing the prior parameters for the parameters of a finite
mixture model.}

\item{\code{type}}{A character specifying what type of prior should be used in
Bayesian estimation. Either \code{"independent"} for an independent prior
distribution or \code{"condconjugate"} for a conditionally conjugate prior
distribution.}

\item{\code{hier}}{A logical defining, if the used prior should be hierarchical.
Hierarchical prior are often more robust, but need an additional layer in
sampling, so computing costs increase.}
}}

\examples{
# Call the default constructor without any arguments.
f_prior <- prior()

}
\references{
\itemize{
\item Fr\"uhwirth-Schnatter, S (2006), "Finite Mixture and Markov Switching Models"
}
}
\seealso{
\itemize{
\item \code{\link[=prior]{prior()}} for the general constructor of this class
\item \code{\link[=priordefine]{priordefine()}} for the advanced constructor of this class
}
}
