% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{priordefine}
\alias{priordefine}
\title{Advanced constructor for the \code{prior} class}
\usage{
priordefine(
  fdata = fdata(),
  model = model(),
  varargin = NULL,
  prior.wagner = TRUE,
  s = 5
)
}
\arguments{
\item{fdata}{An \code{fdata} object holding the data. Observations in slot \verb{@y}
must be existent.}

\item{model}{A \code{model} object specifying the finite mixture model.}

\item{varargin}{\code{NULL} or a \code{prior} object. This enables the user to pass in
an already constructed prior object that gets then completed.}

\item{prior.wagner}{A logical indicating, if the prior from Wagner (2007)
should be used in case of an exponential mixture model.}

\item{s}{A numeric specifying the standard deviation \code{s} for the
Metropolis-Hastings proposal.}
}
\value{
A fully specified \code{prior} object.
}
\description{
This constructor defines a data dependent prior with parameters by matching
moments. As a consequence it needs as inputs an \code{fdata} object and a \code{model}
object. The prior distributions chosen and the methods how parameters are
computed are described in Frühwirth-Schnatter (2006).
}
\examples{
# Create a Poisson mixture model.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Use the advanced constructor to generate a prior.
f_prior <- priordefine(f_data, f_model)

}
\references{
\itemize{
\item Fr\"uwirth-Schnatter, S. (2006), "Finite Mixture and Markov Switching
Models"
\item Wagner, H. (2007), "Bayesian analysis of mixtures of exponentials",
Journal of Applied Mathematics, Statistics and Informatics 3, 165-183
}
}
\seealso{
\itemize{
\item \link[=prior-class]{prior} for the class definition
\item \code{\link[=prior]{prior()}} for the default constructor of the class
}
}
