% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mincol.R
\name{qincolmult}
\alias{qincolmult}
\title{Convert multiple symmetric matrices into vectors}
\usage{
qincolmult(a)
}
\arguments{
\item{q}{A symmetric matrix or dimension \code{rxr}.}
}
\value{
A vector of length \code{r(r+1)/2}.
}
\description{
Calling \code{\link[=qincolmult]{qincolmult()}} on an array of symmetric matrices all with dimension
\code{rxr} converts these matrices into an array of vectors with length
\code{r(r+1)/2}. This function is used to handle the MCMC sampling output from
multivariate finite mixture models. To save storage the symmetric
variance-covariance matrices of multivariate mixtures are stored vector
form. If the covariance matrices are needed for calculations the functions
\code{\link[=qinmatr]{qinmatr()}} and \code{\link[=qinmatrmult]{qinmatrmult()}} helps to restore these matrices from the
storage vectors.
}
\examples{
# Convert a matrix of vectors
matrices <- qinmatrmult(matrix(rnorm(36), nrow = 6))
# Convert these matrices back into vectors.
qincolmult(matrices) 

}
\seealso{
\itemize{
\item \code{\link[=qinmatr]{qinmatr()}} for converting a single vector into a symmetric matrix
\item \code{\link[=qinmatrmult]{qinmatrmult()}} for converting multiple vectors into symmetric matrices
\item \code{\link[=qincol]{qincol()}} for converting a single symmetric matrix into a vector
}
}
