% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mincol.R
\name{qinmatr}
\alias{qinmatr}
\title{Convert vector into matrix.}
\usage{
qinmatr(q)
}
\arguments{
\item{q}{A vector of dimension \verb{r(r+1)/2x1}.}
}
\value{
A symmetric matrix of dimension \code{rxr}.
}
\description{
Calling \code{\link[=qinmatr]{qinmatr()}} on a vector of dimension \verb{r(r+1)/2x1}
converts the vector into a symmetric matrix of dimension \code{rxr}. This
function is used to handle the MCMC sampling output from multivariate finite
mixture models. To save storage the symmetric variance-covariance matrices
of multivariate mixtures are stored vector form. If the covariance matrices
are needed for calculations this function helps to restore these matrices
from the storage vectors.
}
\examples{
# Define a vector.
q <- rnorm(n = 6, mean = 0.5, sd = 2)
# Generate the symmetric matrix.
qinmatr(q)

}
\seealso{
\itemize{
\item \code{\link[=qinmatrmult]{qinmatrmult()}}
\item \code{\link[=qincol]{qincol()}}
\item \code{\link[=qincolmult]{qincolmult()}}
}
}
