% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mincol.R
\name{qinmatrmult}
\alias{qinmatrmult}
\title{Convert array of vectors into array of matrices.}
\usage{
qinmatrmult(m)
}
\arguments{
\item{m}{A matrix or array of vectors of dimension \verb{r(r+1)/2x1}.}
}
\value{
An array of symmetric matrices, all of dimension \code{rxr}.
}
\description{
Calling \code{\link[=qinmatrmult]{qinmatrmult()}} on multiple vectors of dimension \verb{r(r+1)/2x1}
converts these vectors into an array of symmetric matrices of dimension
\code{rxr}. This function is used to handle the MCMC sampling output from
multivariate finite mixture models. To save storage the symmetric
variance-covariance matrices of multivariate mixtures are stored vector
form. If the covariance matrices are needed for calculations this function
helps to restore these matrices from the storage vectors.
}
\examples{
# Convert a matrix of vectors
qinmatrmult(matrix(rnorm(36), nrow = 6))

}
\seealso{
\itemize{
\item \code{\link[=qinmatr]{qinmatr()}} for converting a single vector into a symmetric matrix
\item \code{\link[=qincol]{qincol()}} for converting a symmetric matrix into a vector
\item \code{\link[=qincolmult]{qincolmult()}} for converting an array of symmetric matrices into vectors
}
}
