% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\docType{class}
\name{sdatamoments-class}
\alias{sdatamoments-class}
\alias{.sdatamoments}
\title{Finmix \code{sdatamoments} class}
\description{
Stores moments for indicators of discrete data.
}
\section{Slots}{

\describe{
\item{\code{gmoments}}{A \linkS4class{groupmoments} object storing the
moments for each mixture component.}

\item{\code{fdata}}{An \linkS4class{fdata} object with data from a discrete valued
mixture distribution.}
}}

\seealso{
\itemize{
\item \linkS4class{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
\item \linkS4class{groupmoments} for the parent class
\item \linkS4class{csdatamoments} for the corresponding class defining
moments for data from a continuous-valued finite mixture
}
}
