% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\name{sdatamoments}
\alias{sdatamoments}
\title{Finmix \code{sdatamoments} constructor}
\usage{
sdatamoments(value = fdata())
}
\arguments{
\item{value}{An \linkS4class{fdata} object containing the indicators for
which moments should be calculated.}
}
\value{
If slot \code{type} of the argument \code{value} is \code{"discrete"} an
\code{sdatamoments} object is returned and if the slot is \code{"continuous"},
a \code{csdatamoments} object is returned.
}
\description{
Calling \code{\link[=sdatamoments]{sdatamoments()}} constructs an object of class \code{sdatamoments} or
\code{csdatamoments} depending on the \code{type} slot of the argument \code{value}. If
this slot is \code{"discrete"} an \code{sdatamoments} object is returned and if the
slot is \code{"continuous"}, a \code{csdatamoments} object is returned.
}
\examples{
# Define a model of exponential mixtures.
f_model <- model("exponential", par = list(lambda = c(0.3, 0.7)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Compute data moments for the indicators.
sdatamoments(f_data)

}
\seealso{
\itemize{
\item \linkS4class{sdatamoments} for the class of indicator
moments for discrete data
\item \linkS4class{csdatamoments} for the class of indicator moments
for continuous
\item \linkS4class{groupmoments} for the parent class## Copyright (C) 2013 Lars Simon Zehnder
}
}
