% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\docType{class}
\name{sdatamoments_class}
\alias{sdatamoments_class}
\alias{.sdatamoments}
\alias{show,sdatamoments-method}
\title{Finmix \code{sdatamoments} class}
\usage{
\S4method{show}{sdatamoments}(object)
}
\arguments{
\item{object}{An \code{sdatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Stores moments for indicators of discrete data.

Calling \code{\link[=show]{show()}} on an \code{sdatamoments} object gives an overview
of the moments of a finite mixture with discrete data.
}
\section{Functions}{
\itemize{
\item \code{show,sdatamoments-method}: Shows a summary of an object
}}

\section{Slots}{

\describe{
\item{\code{gmoments}}{A \link[=groupmoments_class]{groupmoments} object storing the
moments for each mixture component.}

\item{\code{fdata}}{An \link[=fdata_class]{fdata} object with data from a discrete valued
mixture distribution.}
}}

\seealso{
\itemize{
\item \link[=datamoments_class]{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
\item \link[=groupmoments_class]{groupmoments} for the parent class
\item \link[=csdatamoments_class]{csdatamoments} for the corresponding class defining
moments for data from a continuous-valued finite mixture
}
}
