% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\docType{class}
\name{sdatamomentsOrNULL-class}
\alias{sdatamomentsOrNULL-class}
\title{Finmix class union of \code{sdatamoments} and \code{NULL}}
\description{
Defines a class union such that the object held by a child class can also
be \code{NULL}.
}
\keyword{internal}
