% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setBycolumn<-,fdata-method}
\alias{setBycolumn<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setBycolumn}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{bycolumn} should be set.}

\item{value}{A logical that should be set as \code{bycolumn} slot of the \code{fdata}
object.}
}
\value{
The \code{fdata} object with slot \code{bycolumn} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{bycolumn}.
}
\description{
Sets the slot \code{bycolumn} of an \code{fdata} object and validates the slot data
before setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setBycolumn(f_data) <- TRUE

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
