% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setBycolumn<-}
\alias{setBycolumn<-}
\title{Setter for the \code{bycolumn} format}
\usage{
setBycolumn(object) <- value
}
\description{
Setter for the \code{bycolumn} format
}
\keyword{internal}
