% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setDist<-,model-method}
\alias{setDist<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setDist}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{A character defining the distribution.}
}
\value{
The \code{model} object with slot \code{dist} set to \code{value}.
}
\description{
Sets a value for the \code{dist} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Get the slot
setDist(f_model) <- "poisson"

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
