% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setIndicfix<-,model-method}
\alias{setIndicfix<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setIndicfix}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{A logical specifying, if the model is one with fixed indicators.}
}
\value{
The \code{model} object with slot \code{indicfix} set to \code{value}.
}
\description{
Sets a value for the \code{indicfix} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the slot.
setIndicfix(f_model) <- TRUE

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
