% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setIndicfix<-}
\alias{setIndicfix<-}
\title{Setter for the \code{indicfix} slot}
\usage{
setIndicfix(object) <- value
}
\description{
Setter for the \code{indicfix} slot
}
\keyword{internal}
