% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setIndicmod<-,model-method}
\alias{setIndicmod<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setIndicmod}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{An character specifying the indicator model.}
}
\value{
The \code{model} object with slot \code{indicmod} set to \code{value}.
}
\description{
Sets a value for the \code{indicmod} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the slot.
setK(f_model) <- 2

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
