% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setIndicmod<-}
\alias{setIndicmod<-}
\title{Setter for the \code{indicmod} slot}
\usage{
setIndicmod(object) <- value
}
\description{
Setter for the \code{indicmod} slot
}
\keyword{internal}
