% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setK<-,model-method}
\alias{setK<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setK}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{An integer specifying the number of components.}
}
\value{
The \code{model} object with slot \code{K} set to \code{value}.
}
\description{
Sets a value for the \code{K} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the slot.
setK(f_model) <- 2

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
