% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setN<-,fdata-method}
\alias{setN<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setN}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{N} should be set.}

\item{value}{An integer that should be set as \code{N} slot of the \code{fdata} object.}
}
\value{
The \code{fdata} object with slot \code{N} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{N}.
}
\description{
Sets the slot \code{N} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
f_data <- fdata()
setN(f_data) <- as.integer(100)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
