% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setName<-,fdata-method}
\alias{setName<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setName}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{name} should be set.}

\item{value}{A matrix that should be set as \code{name} slot of the \code{fdata} object.
Has to be of type integer.}
}
\value{
The \code{fdata} object with slot \code{name} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{name}.
}
\description{
Sets the slot \code{name} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setName(f_data) <- "poisson_data"

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
\keyword{internal}
