% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setPar<-,model-method}
\alias{setPar<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setPar}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{A list specifying the component parameters.}
}
\value{
The \code{model} object with slot \code{par} set to \code{value}.
}
\description{
Sets a value for the \code{par} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the number of components to two.
setK(f_model) <- 2
# Set the slot.
setPar(f_model) <- list(lambda=c(0.2, 0.7))

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
