% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{setPar<-,prior-method}
\alias{setPar<-,prior-method}
\title{Setter method of \code{prior} class.}
\usage{
\S4method{setPar}{prior}(object) <- value
}
\arguments{
\item{object}{An \code{prior} object.}

\item{value}{An integer defining the new value for the \verb{@par} slot.}
}
\value{
None.
}
\description{
Sets the slot. Returns the none.
}
\examples{
# Generate a prior object.
f_prior <- prior()
# Set the slot.
setPar(f_prior) <- setPar(f_prior) <- list(a = matrix(c(1.2, 0.8), nrow = 1), 
                                           b = matrix(c(2.3, 0.4), nrow = 1))
}
\keyword{internal}
