% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setR<-,fdata-method}
\alias{setR<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setR}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{R} should be set.}

\item{value}{An integer that should be set as \code{R} slot of the \code{fdata} object.}
}
\value{
The \code{fdata} object with slot \code{R} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{R}.
}
\description{
Sets the slot \code{R} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
f_data <- fdata()
setR(f_data) <- as.integer(2)

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
\keyword{internal}
