% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setR<-,model-method}
\alias{setR<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setR}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{A character defining the distribution.}
}
\value{
The \code{model} object with slot \code{r} set to \code{value}.
}
\description{
Sets a value for the \code{r} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the slot.
setR(f_model) <- 1

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
