% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{setRanperm<-,mcmc-method}
\alias{setRanperm<-,mcmc-method}
\title{Setter method of \code{mcmc} class.}
\usage{
\S4method{setRanperm}{mcmc}(object) <- value
}
\arguments{
\item{object}{An \code{mcmc} object.}

\item{value}{An integer defining the new value for the \verb{@ranperm} slot.}
}
\value{
None.
}
\description{
Sets the slot. Returns the none.
}
\examples{
# Generate an mcmc object
f_mcmc <- mcmc()
# Set the slot
setRanperm(f_mcmc) <- FALSE

}
\seealso{
\itemize{
\item \linkS4class{mcmc} for the class definition
\item \code{\link[=mcmc]{mcmc()}} for the constructor of the \code{mcmc} class
}
}
\keyword{internal}
