% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setRanperm<-}
\alias{setRanperm<-}
\title{Setter for the \code{ranperm} slot}
\usage{
setRanperm(object) <- value
}
\description{
Setter for the \code{ranperm} slot
}
\keyword{internal}
