% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setS<-,fdata-method}
\alias{setS<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setS}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} object, whose slot \code{S} should be set.}

\item{value}{A matrix that should be set as \code{S} slot of the \code{fdata} object.
Has to be of type integer.}
}
\value{
The \code{fdata} object with slot \code{S} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{S}.
}
\description{
Sets the slot \code{S} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setS(f_data) <- matrix(sample.int(4, 100, replace = TRUE))

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
