% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setSim<-,fdata-method}
\alias{setSim<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setSim}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{sim} should be set.}

\item{value}{A logical that should be set as \code{sim} slot of the \code{fdata} object.
Has to be of type integer.}
}
\value{
The \code{fdata} object with slot \code{sim} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{sim}.
}
\description{
Sets the slot \code{sim} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setSim(f_data) <- TRUE

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
\keyword{internal}
