% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setStartpar<-}
\alias{setStartpar<-}
\title{Setter for the \code{startpar} slot}
\usage{
setStartpar(object) <- value
}
\description{
Setter for the \code{startpar} slot
}
\keyword{internal}
