% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{setStorepost<-}
\alias{setStorepost<-}
\title{Setter for the \code{storepost} slot}
\usage{
setStorepost(object) <- value
}
\description{
Setter for the \code{storepost} slot
}
\keyword{internal}
