% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setT<-,fdata-method}
\alias{setT<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setT}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{T} should be set.}

\item{value}{A matrix that should be set as \code{T} slot of the \code{fdata} object.
Has to be of type integer.}
}
\value{
The \code{fdata} object with slot \code{T} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{T}.
}
\description{
Sets the slot \code{T} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setT(f_data) <- matrix(rep(100, 100))

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
