% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{setT<-,model-method}
\alias{setT<-,model-method}
\title{Setter method of \code{model} class.}
\usage{
\S4method{setT}{model}(object) <- value
}
\arguments{
\item{object}{An \code{model} object.}

\item{value}{An integer specifying the number of components.}
}
\value{
The \code{model} object with slot \code{T} set to \code{value}.
}
\description{
Sets a value for the \code{T} slot.
}
\examples{
# Generate an default mixture model.
f_model <- model()
# Set the slot.
setT(f_model) <- as.integer(4)

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
