% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setType<-,fdata-method}
\alias{setType<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setType}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{type} should be set.}

\item{value}{A character that should be set as \code{type} slot of the \code{fdata} object.
Has to be of type integer.}
}
\value{
The \code{fdata} object with slot \code{type} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{type}.
}
\description{
Sets the slot \code{type} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
# Generate an empty fdata object.
f_data <- fdata()
setType(f_data) <- "discrete"

}
\seealso{
\link{fdata} for all slots of the \code{fdata} class
}
