% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{setType<-,prior-method}
\alias{setType<-,prior-method}
\title{Setter method of \code{prior} class.}
\usage{
\S4method{setType}{prior}(object) <- value
}
\arguments{
\item{object}{An \code{prior} object.}

\item{value}{An integer defining the new value for the \verb{@type} slot.}
}
\value{
None.
}
\description{
Sets the slot. Returns none.
}
\examples{
# Generate a prior object.
f_prior <- prior()
# Set the slot.
setType(f_prior) <- "condconjugate"
}
\keyword{internal}
