% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{setWeight<-,prior-method}
\alias{setWeight<-,prior-method}
\title{Setter method of \code{prior} class.}
\usage{
\S4method{setWeight}{prior}(object) <- value
}
\arguments{
\item{object}{An \code{prior} object.}

\item{value}{An integer defining the new value for the \verb{@weight} slot.}
}
\value{
None.
}
\description{
Sets the slot. Returns the none.
}
\examples{
# Generate a prior object.
f_prior <- prior()
# Set the slot.
setWeight(f_prior) <- matrix(c(0.5, 0.5), nrow = 1)
}
\keyword{internal}
