% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{setY<-,fdata-method}
\alias{setY<-,fdata-method}
\title{Setter method of \code{fdata} class}
\usage{
\S4method{setY}{fdata}(object) <- value
}
\arguments{
\item{object}{An \code{fdata} objects, whose slot \code{y} should be set.}

\item{value}{A matrix that should be set as \code{y} slot of the \code{fdata} object.}
}
\value{
The \code{fdata} object with slot \code{y} set to \code{value} or an error message
if the \code{value} cannot be set as slot \code{y}.
}
\description{
Sets the slot \code{y} of an \code{fdata} object and validates the slot data before
setting.
}
\examples{
f_data <- fdata()
setY(f_data) <- rpois(100, 312)

}
\seealso{
\linkS4class{fdata} for all slots of the \code{fdata} class
}
