% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialmodelmoments.R
\name{show,binomialmodelmoments-method}
\alias{show,binomialmodelmoments-method}
\title{Shows a summary of an \code{binomialmodelmoments} object.}
\usage{
\S4method{show}{binomialmodelmoments}(object)
}
\arguments{
\item{object}{An \code{binomialmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{binomialmodelmoments} object gives an overview
of the moments of an binomial finite mixture.
}
\seealso{
\itemize{
\item \code{\link[=modelmoments]{modelmoments()}} for the mutual constructor for all modelmoments
\item \linkS4class{binomialmodelmoments} for the class definition
}
}
\keyword{internal}
