% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdatamoments.R
\name{show,cdatamoments-method}
\alias{show,cdatamoments-method}
\title{Shows a summary of a \code{cdatamoments} object.}
\usage{
\S4method{show}{cdatamoments}(object)
}
\arguments{
\item{object}{A \code{cdatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on a \code{cdatamoments} object gives an overview
of the moments of a finit mixture with continuous data.
}
