% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdatamoments.R
\name{show,csdatamoments-method}
\alias{show,csdatamoments-method}
\title{Shows a summary of an \code{csdatamoments} object.}
\usage{
\S4method{show}{csdatamoments}(object)
}
\arguments{
\item{object}{An \code{csdatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{csdatamoments} object gives an overview
of the moments of a finite mixture with continuous data.
}
\keyword{internal}
