% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclass.R
\name{show,dataclass-method}
\alias{show,dataclass-method}
\title{Shows a summary of a \code{dataclass} object.}
\usage{
\S4method{show}{dataclass}(object)
}
\arguments{
\item{object}{A \code{dataclass} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on a \code{dataclass} object gives an overview
of the slots of this class.
}
\seealso{
\itemize{
\item \linkS4class{dataclass} for the class definition
\item \code{\link[=dataclass]{dataclass()}} for the class constructor
}
}
\keyword{internal}
