% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddatamoments.R
\name{show,ddatamoments-method}
\alias{show,ddatamoments-method}
\title{Shows a summary of a \code{ddatamoments} object.}
\usage{
\S4method{show}{ddatamoments}(object)
}
\arguments{
\item{object}{A \code{ddatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on a \code{ddatamoments} object gives an overview
of the moments of a finit mixture with continuous data.
}
