% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponentialmodelmoments.R
\name{show,exponentialmodelmoments-method}
\alias{show,exponentialmodelmoments-method}
\title{Shows a summary of an \code{exponentialmodelmoments} object.}
\usage{
\S4method{show}{exponentialmodelmoments}(object)
}
\arguments{
\item{object}{An \code{exponentialmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{exponentialmodelmoments} object gives an overview
of the moments of an exponential finite mixture.
}
