% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{show,fdata-method}
\alias{show,fdata-method}
\title{Shows a summary of an \code{fdata} object.}
\usage{
\S4method{show}{fdata}(object)
}
\arguments{
\item{object}{An \code{fdata} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{fdata} object gives an overview of the different
slots and dimensions.
}
\examples{
# Generate some Poisson data and show the `fdata` object
f_data <- fdata(y = rpois(100, 312), sim = TRUE)
show(f_data)

}
\seealso{
\link{fdata} class for an overview of the slots
}
\keyword{internal}
