% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupmoments.R
\name{show,groupmoments-method}
\alias{show,groupmoments-method}
\title{Shows a summary of a \code{groupmoments} object.}
\usage{
\S4method{show}{groupmoments}(object)
}
\arguments{
\item{object}{A \code{groupmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on a \code{groupmoments} object gives an overview
of the moments of a finit mixture with continuous data.
}
\keyword{internal}
