% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{show,mcmc-method}
\alias{show,mcmc-method}
\title{Shows a summary of an \code{mcmc} object.}
\usage{
\S4method{show}{mcmc}(object)
}
\arguments{
\item{object}{A \code{mcmc} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmc} object gives an overview
of the \code{mcmc} object.
}
