% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestfix.R
\name{show,mcmcestfix-method}
\alias{show,mcmcestfix-method}
\title{Shows a summary of an \code{mcmcestfix} object.}
\usage{
\S4method{show}{mcmcestfix}(object)
}
\arguments{
\item{object}{An \code{mcmcestfix} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcestfix} object gives an overview
of the \code{mcmcestfix} object.
}
