% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestind.R
\name{show,mcmcestind-method}
\alias{show,mcmcestind-method}
\title{Shows a summary of an \code{mcmcestind} object.}
\usage{
\S4method{show}{mcmcestind}(object)
}
\arguments{
\item{object}{An \code{mcmcestind} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcestind} object gives an overview
of the \code{mcmcestind} object.
}
\keyword{internal}
