% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputbase.R
\name{show,mcmcoutputbase-method}
\alias{show,mcmcoutputbase-method}
\title{Shows a summary of an \code{mcmcoutputbase} object.}
\usage{
\S4method{show}{mcmcoutputbase}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputbase} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputbase} object gives an overview
of the \code{mcmcoutputbase} object.
}
