% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfix.R
\name{show,mcmcoutputfix-method}
\alias{show,mcmcoutputfix-method}
\title{Shows a summary of an \code{mcmcoutputfix} object.}
\usage{
\S4method{show}{mcmcoutputfix}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputfix} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputfix} object gives an overview
of the \code{mcmcoutputfix} object.
}
\keyword{internal}
