% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhier.R
\name{show,mcmcoutputfixhier-method}
\alias{show,mcmcoutputfixhier-method}
\title{Shows a summary of an \code{mcmcoutputfixhier} object.}
\usage{
\S4method{show}{mcmcoutputfixhier}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputfixhier} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputfixhier} object gives an overview
of the \code{mcmcoutputfixhier} object.
}
