% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhierpost.R
\name{show,mcmcoutputfixhierpost-method}
\alias{show,mcmcoutputfixhierpost-method}
\title{Shows a summary of an \code{mcmcoutputfixhierpost} object.}
\usage{
\S4method{show}{mcmcoutputfixhierpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputfixhierpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputfixhierpost} object gives an overview
of the \code{mcmcoutputfixhierpost} object.
}
