% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixpost.R
\name{show,mcmcoutputfixpost-method}
\alias{show,mcmcoutputfixpost-method}
\title{Shows a summary of an \code{mcmcoutputfixpost} object.}
\usage{
\S4method{show}{mcmcoutputfixpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputfixpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputfixpost} object gives an overview
of the \code{mcmcoutputfixpost} object.
}
\keyword{internal}
