% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhier.R
\name{show,mcmcoutputhier-method}
\alias{show,mcmcoutputhier-method}
\title{Shows a summary of an \code{mcmcoutputhier} object.}
\usage{
\S4method{show}{mcmcoutputhier}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputhier} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputhier} object gives an overview
of the \code{mcmcoutputhier} object.
}
\keyword{internal}
