% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhierpost.R
\name{show,mcmcoutputhierpost-method}
\alias{show,mcmcoutputhierpost-method}
\title{Shows a summary of an \code{mcmcoutputhierpost} object.}
\usage{
\S4method{show}{mcmcoutputhierpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputhierpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputhierpost} object gives an overview
of the \code{mcmcoutputhierpost} object.
}
\keyword{internal}
