% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermbase.R
\name{show,mcmcoutputpermbase-method}
\alias{show,mcmcoutputpermbase-method}
\title{Shows a summary of an \code{mcmcoutputpermbase} object.}
\usage{
\S4method{show}{mcmcoutputpermbase}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermbase} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermbase} object gives an overview
of the \code{mcmcoutputpermbase} object.
}
\keyword{internal}
