% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfix.R
\name{show,mcmcoutputpermfix-method}
\alias{show,mcmcoutputpermfix-method}
\title{Shows a summary of an \code{mcmcoutputpermfix} object.}
\usage{
\S4method{show}{mcmcoutputpermfix}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermfix} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermfix} object gives an overview
of the \code{mcmcoutputpermfix} object.
}
