% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixhier.R
\name{show,mcmcoutputpermfixhier-method}
\alias{show,mcmcoutputpermfixhier-method}
\title{Shows a summary of an \code{mcmcoutputpermfixhier} object.}
\usage{
\S4method{show}{mcmcoutputpermfixhier}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermfixhier} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermfixhier} object gives an overview
of the \code{mcmcoutputpermfixhier} object.
}
\keyword{internal}
