% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixhierpost.R
\name{show,mcmcoutputpermfixhierpost-method}
\alias{show,mcmcoutputpermfixhierpost-method}
\title{Shows a summary of an \code{mcmcoutputpermfixhierpost} object.}
\usage{
\S4method{show}{mcmcoutputpermfixhierpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermfixhierpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermfixhierpost} object gives an overview
of the \code{mcmcoutputpermfixhierpost} object.
}
\keyword{internal}
