% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixpost.R
\name{show,mcmcoutputpermfixpost-method}
\alias{show,mcmcoutputpermfixpost-method}
\title{Shows a summary of an \code{mcmcoutputpermfixpost} object.}
\usage{
\S4method{show}{mcmcoutputpermfixpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermfixpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermfixpost} object gives an overview
of the \code{mcmcoutputpermfixpost} object.
}
