% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhier.R
\name{show,mcmcoutputpermhier-method}
\alias{show,mcmcoutputpermhier-method}
\title{Shows a summary of an \code{mcmcoutputpermhier} object.}
\usage{
\S4method{show}{mcmcoutputpermhier}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermhier} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermhier} object gives an overview
of the \code{mcmcoutputpermhier} object.
}
