% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhierpost.R
\name{show,mcmcoutputpermhierpost-method}
\alias{show,mcmcoutputpermhierpost-method}
\title{Shows a summary of an \code{mcmcoutputpermhierpost} object.}
\usage{
\S4method{show}{mcmcoutputpermhierpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermhierpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermhierpost} object gives an overview
of the \code{mcmcoutputpermhierpost} object.
}
\keyword{internal}
