% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermpost.R
\name{show,mcmcoutputpermpost-method}
\alias{show,mcmcoutputpermpost-method}
\title{Shows a summary of an \code{mcmcoutputpermpost} object.}
\usage{
\S4method{show}{mcmcoutputpermpost}(object)
}
\arguments{
\item{object}{An \code{mcmcoutputpermpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{mcmcoutputpermpost} object gives an overview
of the \code{mcmcoutputpermpost} object.
}
\keyword{internal}
