% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpost.R
\name{show,mcmcoutputpost-method}
\alias{show,mcmcoutputpost-method}
\title{Shows a summary of an \code{ mcmcoutputpost} object.}
\usage{
\S4method{show}{mcmcoutputpost}(object)
}
\arguments{
\item{object}{An \code{ mcmcoutputpost} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{ mcmcoutputpost} object gives an overview
of the \code{ mcmcoutputpost} object.
}
\keyword{internal}
