% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{show,model-method}
\alias{show,model-method}
\title{Shows the model.}
\usage{
\S4method{show}{model}(object)
}
\arguments{
\item{object}{An S4 model object.}
}
\value{
A print out of model information about all slots.
}
\description{
\code{show} prints model information to the console.
}
\examples{
\dontrun{
show(f_model)
}

}
\seealso{
\itemize{
\item \linkS4class{model} for the class definition
}
}
\keyword{internal}
