% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normultmodelmoments.R
\name{show,normultmodelmoments-method}
\alias{show,normultmodelmoments-method}
\title{Shows a summary of an \code{normultmodelmoments} object.}
\usage{
\S4method{show}{normultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normultmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{normultmodelmoments} object gives an overview
of the moments of an normult finite mixture.
}
