% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{show,prior-method}
\alias{show,prior-method}
\title{Shows a summary of a \code{prior} object.}
\usage{
\S4method{show}{prior}(object)
}
\arguments{
\item{object}{A \code{prior} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on a \code{prior} object gives an overview
of the slots of a \code{prior} object.
}
\seealso{
\itemize{
\item \linkS4class{prior} for the class definition
\item \code{\link[=prior]{prior()}} for the basic constructor of the class
\item \code{\link[=priordefine]{priordefine()}} for the advanced constructor of the class
}
}
\keyword{internal}
