% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentmodelmoments.R
\name{show,studentmodelmoments-method}
\alias{show,studentmodelmoments-method}
\title{Shows a summary of an \code{studentmodelmoments} object.}
\usage{
\S4method{show}{studentmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studentmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{studentmodelmoments} object gives an overview
of the moments of an student finite mixture.
}
