% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stephens1997a_binomial_cc}
\alias{stephens1997a_binomial_cc}
\title{Stephens (1997a) relabeling algorithm for Binomial mixtures}
\usage{
stephens1997a_binomial_cc(values1, values2, pars, perm)
}
\arguments{
\item{values1}{A matrix containing the sampled component parameters \code{p}.}

\item{values2}{A matrix containing the sampled weights.}

\item{pars}{A vector containing the parameters of the posterior
distributions of the component parameters and weights. More specifically,
the parameters of the Dirichlet distribution for the weights and the shape
and rate parameters for the Beta distributions of the component
parameters.}

\item{perm}{A matrix with all possible permutations of the labels.}
}
\value{
A matrix of dimension \code{MxK} that holding the optimal labeling.
}
\description{
For internal usage only. This function runs the re-labeling
algorithm from Stephens (1997a) for MCMC samples of a Binomial mixture
distribution. For optimization a Nelder-Mead-Algorithm from the NLopt
library is used. This is also the reason why the package depends on the
\code{nloptr} package which provides a header file for direct access to the C
routines.
}
\references{
\itemize{
\item Stephens, Matthew (1997a), Discussion of "On Bayesian Analysis of
Mixtures with an Unknown Number of Components",  Journal of the Royal
Statistical Society: Series B (Statistical Methodology), 59: 731-792.
}
}
\seealso{
\itemize{
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for the calling function
\item \code{\link[=stephens1997b_poisson_cc]{stephens1997b_poisson_cc()}} for the re-labeling algorithm from
Stephens (1997b)
\item \code{\link[=stephens1997a_binomial_cc]{stephens1997a_binomial_cc()}} for the equivalent implementation
for mixtures of Binomial distributions
}
}
\keyword{internal}
