% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stephens1997b_poisson_cc}
\alias{stephens1997b_poisson_cc}
\title{Stephens (1997b) relabeling algorithm for Poisson mixtures}
\usage{
stephens1997b_poisson_cc(values, comp_par, weight_par, max_iter = 200L)
}
\arguments{
\item{values}{A matrix of observations of dimension \code{Nx1}.}

\item{comp_par}{An array of component parameter samples from MCMC sampling.
Dimension is \code{MxK}.}

\item{max_iter}{A signed integer specifying the number of iterations to be
run in optimization. Unused.}

\item{weight}{An array of weight parameter samples from MCMC sampling.
Dimension is \code{MxK}.}
}
\value{
An integer matrix of dimension \code{MxK} that holding the optimal
labeling.
}
\description{
For internal usage only. This function runs the re-labeling algorithm from
Stephens (1997b) for MCMC samples of a Poisson mixture distribution.
}
\references{
\itemize{
\item Stephens, Matthew (1997a), Bayesian methods for mixtures of normal
distributions, DPhil Thesis, University of Oxford, Oxford.
}
}
\seealso{
\itemize{
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for the calling function
\item \code{\link[=stephens1997a_poisson_cc]{stephens1997a_poisson_cc()}} for the re-labeling algorithm from
Stephens (1997a)
\item \code{\link[=stephens1997b_binomial_cc]{stephens1997b_binomial_cc()}} for the equivalent implementation
for mixtures of Binomial distributions
}
}
\keyword{internal}
