% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studmultmodelmoments.R
\docType{class}
\name{studmultmodelmoments}
\alias{studmultmodelmoments}
\alias{.studmultmodelmoments}
\title{Finmix \code{studmultmodelmoments} class}
\description{
Defines a class that holds modelmoments for a finite mixture of studmult
distributions. Note that this class is not directly used, but indirectly
when calling the \code{modelmoments} constructor \code{\link[=modelmoments]{modelmoments()}}.
}
\section{Slots}{

\describe{
\item{\code{B}}{A numeric defining the between-group heterogeneity.}

\item{\code{W}}{A numeric defining the within-group heterogeneity.}

\item{\code{R}}{A numeric defining the coefficient of determination.}
}}

\seealso{
\itemize{
\item \linkS4class{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
}
}
