% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{subseq-generic}
\alias{subseq-generic}
\title{Generates a sub-chain from MCMC samples}
\usage{
subseq(object, index)
}
\description{
Generates a sub-chain from MCMC samples
}
\keyword{internal}
