% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhier.R
\name{subseq,mcmcoutputhier,array-method}
\alias{subseq,mcmcoutputhier,array-method}
\title{Constructs a sub-chain of MCMC samples}
\usage{
\S4method{subseq}{mcmcoutputhier,array}(object, index)
}
\arguments{
\item{object}{An \code{mcmcoutput} object containing all sampled values.}

\item{index}{An array specifying the extraction of the sub-chain.}
}
\value{
An \code{mcmcoutput} object containing the values from the sub-chain.
}
\description{
Calling \code{\link[=subseq]{subseq()}} constructs an MCMC sub-chain from the samples in the
passed-in \code{mcmcoutput} object specfied by the index \code{array} in \code{index}. This
can be advantageous, if chains are non-stationary. For successful MCMC
sampling the chain must be converged to the target distribution, the true
distribution of parameters, weights and indicators.

Note, this method calls the equivalent method of the parent class and then
adds to it the sub-chains for the parameters of the hierarchical prior.
}
\keyword{internal}
