% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhierpost.R
\name{subseq}
\alias{subseq}
\title{Extract sub-chains from MCMC samples}
\arguments{
\item{object}{An \code{mcmcoutput} or \code{mcmcoutputperm} object containing samples
from MCMC samples.}

\item{index}{A logical \code{array} of dimension \code{Mx1} defining the schema for
the sub-chain.}
}
\value{
An \code{mcmcoutput} or \code{mcmcoutputperm} object containing the
sub-chained MCMC samples.
}
\description{
Calling \code{\link[=subseq]{subseq()}} on an \code{mcmcoutput} or \code{mcmcoutputperm} object creates a
sub-chain defined by the argument \code{index}. Sub-chains can be used to further
investigate convergence of MCMC sampling.
}
\details{
Running MCMC sampling should by time result in a roughly stationary sequence
of random draws. If trace plots do not show this stationary pattern MCMC
sampling should be run with a longer burn-in period until the sampling
distribution has converged. Another possibility is to remove the first draws.
Removing the first draws can be achieved by calling \code{subseq()} on the object
holding the MCMC samples.
In case of autocorrelations in the traces it is also possible to extract
every \code{t}-th value by setting the \code{index} argument accordingly.
}
\examples{
# Define a mixture of Poisson distributions.
f_model <- model("poisson", par = list(lambda = c(0.3, 1.2)), K = 2)
# Simulate data from the model.
f_data <- simulate(f_model)
# Define the hyper-parameters for MCMC sampling.
f_mcmc <- mcmc(storepost = FALSE)
# Define the prior distribution by relying on the data.
f_prior <- priordefine(f_data, f_model)
# Start MCMC sampling.
f_output <- mixturemcmc(f_data, f_model, f_prior, f_mcmc)
# Define a sub-chain randomly.
index <- array(sample(c(FALSE, TRUE), size = getM(f_output), replace = TRUE))
# Extract the sub-chain.
subseq(f_output, index)

}
\seealso{
\itemize{
\item \linkS4class{mcmcoutput} for the class storing MCMC samples
\item \linkS4class{mcmcoutputperm} for the corresponding class for re-labeled MCMC
samples
\item \code{\link[=plotTraces]{plotTraces()}} for plotting traces to be used for a convergence analysis
\item \code{\link[=swapElements]{swapElements()}} for swapping elements in MCMC samples
}
}
