% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestfix.R
\name{Summary,mcmcestfix-method}
\alias{Summary,mcmcestfix-method}
\title{Shows an advanced summary of an \code{mcmcestfix} object.}
\usage{
\S4method{Summary}{mcmcestfix}(x, ..., na.rm = FALSE)
}
\arguments{
\item{object}{An \code{mcmcestfix} object.}
}
\value{
A console output listing the formatted slots and summary
information about each of them.
}
\description{
Calling \code{\link[=Summary]{Summary()}} on an \code{mcmcestfix} object gives an advanced overview
of the \code{mcmcestfix} object.
}
\details{
Note, this method is so far only implemented for mixtures of Poisson
distributions.
}
\keyword{internal}
