% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcestind.R
\name{Summary,mcmcestind-method}
\alias{Summary,mcmcestind-method}
\title{Shows an advanced summary of an \code{mcmcestind} object.}
\usage{
\S4method{Summary}{mcmcestind}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{mcmcestind} object.}

\item{...}{(Not used) Further arguments to pass to called functions.}

\item{na.rm}{(Not used) A logical indicating, if \code{NA} values should be
removed in calculating statistics.}
}
\value{
A console output listing the formatted slots and summary
information about each of them.
}
\description{
Calling \code{\link[=Summary]{Summary()}} on an \code{mcmcestind} object gives an advanced overview
of the \code{mcmcestind} object.
}
\details{
Note, this method is so far only implemented for mixtures of Poisson
distributions.
}
