% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_3d_cc}
\alias{swap_3d_cc}
\title{Swap elements in a 3d array}
\usage{
swap_3d_cc(values, index)
}
\arguments{
\item{values}{An array of dimension \verb{M x r x K} of values to swap.}

\item{index}{An integer matrix of dimension \verb{M x K}. containing the scheme
by which values should be swapped.}

\item{A}{three-dimensional array with swapped values.}
}
\description{
This function swaps the elements in a three-dimensional array by using the
scheme provided in the \code{index} matrix.
}
\examples{
values <- array(rnorm(40), dim = c(10, 2, 2))
index <- matrix(c(1,2), nrow = 10, ncol = 2)
swap_3d_cc(values, index)

}
\seealso{
\itemize{
\item \link[=mcmcoutput_class]{swapElements()} for the calling method
\item \code{\link[=swap_cc]{swap_cc()}} for the equivalent function for 2-dimensional arrays
}
}
