% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swap_cc}
\alias{swap_cc}
\title{Swaps values in each row}
\usage{
swap_cc(values, index)
}
\arguments{
\item{values}{A matrix containing the values to be swapped.}

\item{index}{An integer matrix defining how values should be swapped.}
}
\value{
A matrix with swapped values.
}
\description{
This function swaps the values in each row of a matrix by permuting the
columns via the indices provided in the \code{index} matrix. All
\code{swapElements()}-methods use this function internally. The code is extended
to \verb{C++} using the packages \code{Rcpp} and \code{RcppArmadillo}.
}
\examples{
values <- matrix(rnorm(10), nrow = 2)
index <- matrix(c(2,1), nrow = 5, ncol = 2)
swap_cc(values, index)

}
\seealso{
\itemize{
\item \link[=mcmcoutput_class]{swapElements()} for the calling function
}
}
