% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputfixhierpost.R
\name{swapElements,mcmcoutputfixhierpost,array-method}
\alias{swapElements,mcmcoutputfixhierpost,array-method}
\title{Swaps elements between components}
\usage{
\S4method{swapElements}{mcmcoutputfixhierpost,array}(object, index)
}
\arguments{
\item{object}{An \code{mcmcoutput} object containing the sampled values.}

\item{index}{An array specifying the extraction of the values.}
}
\value{
An \code{mcmcoutput} object with swapped elements.
}
\description{
Not yet implemented.
}
\keyword{internal}
