% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhierpost.R
\name{swapElements,mcmcoutputhierpost,array-method}
\alias{swapElements,mcmcoutputhierpost,array-method}
\title{Swaps elements between components}
\usage{
\S4method{swapElements}{mcmcoutputhierpost,array}(object, index)
}
\arguments{
\item{object}{An \code{mcmcoutputhierpost} object containing the sampled values.}

\item{index}{An array specifying the extraction of the values.}
}
\value{
An \code{mcmcoutputhierpost} object with swapped elements.
}
\description{
Not yet implemented.
}
\keyword{internal}
