% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpost.R
\name{swapElements,mcmcoutputpost,array-method}
\alias{swapElements,mcmcoutputpost,array-method}
\title{Swaps elements between components}
\usage{
\S4method{swapElements}{mcmcoutputpost,array}(object, index)
}
\arguments{
\item{object}{An \code{mcmcoutputpost} object containing the sampled values.}

\item{index}{An array specifying the extraction of the values.}
}
\value{
An \code{mcmcoutputpost} object with swapped elements.
}
\description{
Not yet implemented.
}
\keyword{internal}
