% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swapInd_cc}
\alias{swapInd_cc}
\title{Swap values of stored indicators}
\usage{
swapInd_cc(values, index)
}
\arguments{
\item{values}{An integer matrix containing the last indicators stored in
MCMC sampling. The number of these last stored indicators is defined by
the hpyer-parameter \code{storeS} in the \code{mcmc} object.}

\item{index}{An integer matrix defining the swapping scheme.}
}
\value{
A matrix with swapped values.
}
\description{
This function is used to swap elements in the stored indicators from MCMC
sampling. Note that this function reuses R memory and should therefore be
treated with caution. Do not use this function unless you really know what
you are doing.
}
\seealso{
\itemize{
\item \code{\link[=mcmc]{mcmc()}} for the hyper-parameter \code{storeS}
\item \code{\link[=swapElements]{swapElements()}} for the calling method
\item \code{\link[=swapInteger_cc]{swapInteger_cc()}} for the equivalent function that swaps simple integer
matrices
\item \code{\link[=swap_3d_cc]{swap_3d_cc()}} for a function that swaps values in three-dimensional
arrays
}
}
