% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swapInteger_cc}
\alias{swapInteger_cc}
\title{Swap values in an integer matrix}
\usage{
swapInteger_cc(values, index)
}
\arguments{
\item{values}{An integer matrix containing the values to swap.}

\item{index}{An integer matrix containing the indices by which values
should be swapped.}
}
\value{
An integer matrix containing the swapped values.
}
\description{
This function swaps the values in an integer matrix column-wise defined
by the \code{index} matrix. This function is used mainly for the
\code{swapElements()}-method of MCMC samples to swap the indicator values.
}
\examples{
values <- matrix(c(2, 4, 1, 3), nrow = 10, ncol = 2)
index <- matrix(c(1, 2), nrow = 10, ncol = 2)
swapInteger_cc(values, index)

}
\seealso{
\itemize{
\item \code{\link[=swap_cc]{swap_cc()}} for the equivalent function for numeric values
\item \code{\link[=swap_3d_cc]{swap_3d_cc()}} for the equivalent function for three-dimensional arrays
}
}
