% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{swapST_cc}
\alias{swapST_cc}
\title{Swap the \code{ST} slot in the MCMC output}
\usage{
swapST_cc(values, index)
}
\arguments{
\item{values}{An integer matrix containing the values to swap in R memory.}

\item{index}{An integer matrix containing the swapping scheme.}
}
\value{
An integer matrix with swapped values.
}
\description{
This function is used to swap the elements in slot \code{ST} of an \code{mcmcoutput}
object (An MCMC sampling output). The main difference to the
\code{\link[=swapInteger_cc]{swapInteger_cc()}} function is that this function reuses memory from R. Do
only use this function, if you really know what you are doing.
}
\seealso{
\itemize{
\item \code{\link[=swapInteger_cc]{swapInteger_cc()}} for the equivalent function not using R memory
\item \code{\link[=swap_3d_cc]{swap_3d_cc()}} for an equivalent function for three-dimensional arrays
\item \code{\link[=swapElements]{swapElements()}} for the calling method
}
}
