% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unass.R
\name{unsass}
\alias{unsass}
\alias{\%=\%}
\title{Unstructuring assignments}
\usage{
unsass(lhs, rhs)
}
\arguments{
\item{lhs}{A \code{formula} chaining multiple objects (names) together by
\code{~}. These are the objects (names) the right-hand side should be
assigned to.}

\item{rhs}{A \code{list} of objects that should be assigned to the left-hand
side \code{lhs}.}
}
\description{
\code{unsass} assigns multiple objects in its argument
\code{rhs} (right-hand side) to multiple objects (names) chained in its
argument \code{lhs} (left-hand-side).

This is a helper function to simplify the use of the package. The right-hand
side can be a function that returns multiple objects and the left-hand side
must be a formula with objects (names) chained by \code{~}. Assignment works
via \code{lhs \%=\% rhs}.
}
\examples{
f_model <- model(K=2, dist= 'poisson', par=list(lambda=c(0.17, 0.12))) 
f_data <- simulate(f_model) 
mcmc <- mcmc() 
(f_data~f_model~mcmc) \%=\% mcmcstart(f_data, f_model, mcmc)

}
\seealso{
\itemize{
\item \code{\link{mcmcstart}} for generating starting parameters or indicators
}

\itemize{
\item \code{\link{unsass}}
}
}
\author{
Barry Rowlingson (January, 2013)
}
