% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialmodelmoments.R
\docType{class}
\name{binomialmodelmoments-class}
\alias{binomialmodelmoments-class}
\alias{.binomialmodelmoments}
\title{Finmix \code{binomialmodelmoments} class}
\description{
Defines a class that holds modelmoments for a finite mixture of Binomial
distributions. Note that this class is not directly used, but indirectly
when calling the \code{modelmoments} constructor \code{\link[=modelmoments]{modelmoments()}}.

This is a class that directly inherits from the \code{dmodelmoments} class.
}
\seealso{
\itemize{
\item \linkS4class{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
\item \linkS4class{dmodelmoments} class for the parent class
}
}
\keyword{internal}
