% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdatamoments.R
\docType{class}
\name{csdatamoments-class}
\alias{csdatamoments-class}
\alias{.csdatamoments}
\alias{show,csdatamoments-method}
\title{Finmix \code{csdatamoments} class}
\usage{
\S4method{show}{csdatamoments}(object)
}
\arguments{
\item{object}{An \code{csdatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Stores moments for indicators of continuous data. Inherited directly from
the \linkS4class{sdatamoments} class.

Calling \code{\link[=show]{show()}} on an \code{csdatamoments} object gives an overview
of the moments of a finite mixture with continuous data.
}
\section{Functions}{
\itemize{
\item \code{show,csdatamoments-method}: Shows a short summary of the object's slots.
}}

\section{Slots}{

\describe{
\item{\code{B}}{A vector storing the between-group heterogeneity.}

\item{\code{W}}{A vector storing the within-group heterogeneity.}

\item{\code{T}}{A vector storing the total variance.}

\item{\code{R}}{A numeric storing the coefficient of determination for univariate
data.}

\item{\code{Rdet}}{A numeric storing the coefficient of determination using the
trace for multivariate data.}

\item{\code{Rtr}}{A numeric storing the coefficient of determination using the
determinants for multivariate data.}
}}

\seealso{
\itemize{
\item \linkS4class{datamoments} for the base class for data moments
\item \code{\link[=datamoments]{datamoments()}} for the constructor of any object of the \code{datamoments}
class family
\item \linkS4class{csdatamoments} for the corresponding class defining
moments for data from a discrete-valued finite mixture
}
}
\keyword{internal}
