% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclass.R
\name{dataclass}
\alias{dataclass}
\title{Finmix \code{dataclass} constructor}
\usage{
dataclass(fdata = NULL, model = NULL, simS = FALSE)
}
\arguments{
\item{fdata}{An \code{fdata} object containing observations in slot \verb{@y} and
indicators in slot \verb{@S}.}

\item{model}{A \code{model} object containing parameters in slot  \verb{@par} and
and weights in slot \verb{@weight}.}

\item{simS}{A logical defining, if the indicators \code{S} should be simulated.}
}
\value{
A \code{dataclass} object containing the classification matrix,
model log-likelihood, entropy and indicators, if the latter have been
simulated.
}
\description{
Calling \code{\link[=dataclass]{dataclass()}} classifies data using a fully specified mixture model.
Henceforth, the finite mixture model \code{model} must be fully specified, i.e.
containing parameters in slot \verb{@par}, weights in slot \verb{@weight} and
indicators in slot \verb{@S} of the corresponding \code{fdata} object.
}
\seealso{
\itemize{
\item \link[=dataclass_class]{dataclass} for the class definition
}

#' @references
Frühwirth-Schnatter, S. (2006), "Finite Mixture and Markov Switching Models"
}
