% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamoments.R
\name{datamoments}
\alias{datamoments}
\title{Constructor for \code{datamoments} classes}
\usage{
datamoments(value = fdata())
}
\arguments{
\item{value}{An \code{fdata} object with at least slot \code{y} non-empty.}
}
\value{
An \code{datamoments} object containing the data moments for slot \code{y}
and if available slot \code{S}.
}
\description{
Calling \code{\link[=datamoments]{datamoments()}} generates the datamoments for an \code{fdata} object.
Depending on the type of data either an \code{cdatamoments} or \code{ddatamoments}
object is generated. If in addition the \code{fdata} object containes fixed
indicators, these \code{datamoments} object also hold an \code{sdatamoments} class to
store the data moments of these indicators.
}
\examples{
# Create an fdata class with Poisson data.
f_data <- fdata(rpois(100, 312), sim=TRUE)
# Compute the data moments.
datamoments(f_data)

}
\seealso{
\itemize{
\item \link{datamoments} class for all slots of this class
\item \link{cdatamoments} for the class for continuous data
\item \link{ddatamoments} for the class for discrete data
}
}
