% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdata.R
\name{fdata}
\alias{fdata}
\title{Constructs an \code{fdata} object}
\usage{
fdata(
  y = matrix(),
  N = 1,
  r = 1,
  S = matrix(),
  bycolumn = TRUE,
  name = character(),
  type = "discrete",
  sim = FALSE,
  exp = matrix(),
  T = matrix()
)
}
\arguments{
\item{y}{A matrix containing the observations for finite mixture estimation.
Can be by column or row depending on the slot \code{bycolumn}.}

\item{N}{An integer holding the number of observations.}

\item{r}{An integer defining the dimension of the data. Only for multivariate
distributions like \code{normult} or \code{studmult} the dimension is
larger one.}

\item{S}{A matrix containing the indicators of the data. If the \code{fdata} class
contains indicators estimation is performed with a fixed indicator
approach.}

\item{bycolumn}{A logical indicating if the data in \code{y} and \code{S} is sorted by
by column (\code{TRUE}) or row (\code{FALSE}).}

\item{name}{A character specifying a name for the data. Optional.}

\item{type}{A character specifying the data type: either \code{discrete} for
discrete data or \code{continuous} for continuous data. The two data types are
treated differently when calculating data moments.}

\item{sim}{A logical indicating, if the data was simulated.}

\item{exp}{A matrix containing the \emph{exposures} of Poisson data.}

\item{T}{A matrix containing the (optional) repetitions of binomial or Poisson
data. Must be of type integer.}
}
\description{
Calling \code{\link[=fdata]{fdata()}} constructs an \code{fdata} object. Can be called without
arguments.
}
\examples{
# Call the constructor without arguments.
f_data <- fdata()

# Create simulated data.
f_data <- fdata(y = rpois(100, 312), sim = TRUE)

}
\seealso{
\link{fdata} class that describes the slots and the getters, setters and
and checkers
}
