% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputbase.R
\docType{class}
\name{mcmcoutputbase-class}
\alias{mcmcoutputbase-class}
\alias{.mcmcoutputbase}
\title{Finmix \code{mcmcoutput} base class for unknown indicators}
\description{
This class defines the basic slots for the MCMC sampling output when
indicators are not known. It inherits from the
\linkS4class{mcmcoutputfix}.
}
\section{Slots}{

\describe{
\item{\code{weight}}{An \code{array} of dimension \verb{M x K} containing the sampled
weights.}

\item{\code{entropy}}{An \code{array} of dimension \verb{M x 1} containing the entropy
for each MCMC draw.}

\item{\code{ST}}{An \code{array} of dimension \verb{M x 1} containing all MCMC states,
for the last observation in slot \code{y} of the \code{fdata} object passed in to
\code{\link[=mixturemcmc]{mixturemcmc()}} where a state is defined for non-Markov models as the
last indicator of this observation.}

\item{\code{S}}{An \code{array} of dimension \verb{N x storeS} containing the last
\code{storeS} indicators sampled. \code{storeS} is defined in the slot \code{storeS} of
the \code{mcmc} object passed into \code{\link[=mixturemcmc]{mixturemcmc()}}.}

\item{\code{NK}}{An \code{array} of dimension \verb{M x K} containing the number of
observations assigned to each component for each MCMC draw.}

\item{\code{clust}}{An \code{array} of dimension \verb{N x 1} containing the recent
indicators defining the last "clustering" of observations into the
mixture components.}
}}

