% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputhierpost.R
\docType{class}
\name{mcmcoutputhierpost-class}
\alias{mcmcoutputhierpost-class}
\alias{.mcmcoutputhierpost}
\title{Finmix \code{mcmcoutputhierpost} class}
\description{
This class stores samples from bayesian estimation with hierarchical prior
and unknown indicators. It inherits from \code{mcmcoutputhier} and adds to it a
slot to store the parameters from the posterior density. For a model with
unknown indicators the slot \verb{@indicfix} in the \code{model} object specifying
the finite mixture model must be set to \code{FALSE} (default). Sampling with a
hierarchical prior is activated by setting the slot \verb{@hier} in the \code{prior}
object to \code{TRUE} (default). Finally, to store parameters for the posterior
density the hyper-parameter \code{storepost} in the \code{mcmc} object must be set to
\code{TRUE} (default).
}
\section{Slots}{

\describe{
\item{\code{post}}{A named list containing a named list \code{par} that contains arrays
storing the sampled posterior density parameters.}
}}

\seealso{
\itemize{
\item \linkS4class{mcmcoutputhier} for the parent class
\item \linkS4class{prior} for the class specifying the prior distribution
\item \code{\link[=prior]{prior()}} for the \code{prior} class constructor
\item \code{\link[=priordefine]{priordefine()}} for the advanced \code{prior} class constructor
\item \linkS4class{mcmc} for the class defining the hyper-parameters
\item \code{\link[=mcmc]{mcmc()}} for the \code{mcmc} class constructor
\item \linkS4class{model} for the \code{model} class definition
\item \code{\link[=model]{model()}} for the \code{model} class constructor
}
}
