% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermhier.R
\docType{class}
\name{mcmcoutputpermhier-class}
\alias{mcmcoutputpermhier-class}
\alias{.mcmcoutputpermhier}
\title{Finmix \code{mcmcoutputpermhier} class}
\description{
This class defines objects to store the outputs from permuting the MCMC
samples. Due to label switching the sampled component parameters are usually
not assigned to the same component in each iteration. To overcome this issue
the samples are permuted by using a relabeling algorithm (usually K-means)
to reassign parameters. Note that due to assignment of parameters from the
same iteration to the same component, the sample size could shrink.

This class stores the permuted parameters together with the new MCMC sample
size and the mixture log-likelihood, the prior log-likelihood, and the
complete data posterior log-likelihood.

Note that this class inherits all of its slots from the parent classes.
}
\seealso{
\itemize{
\item \linkS4class{mcmcoutputhier} for the parent class
\item \linkS4class{mcmcpermindhier} for the parent class
\item \code{\link[=mcmcpermute]{mcmcpermute()}} for performing permutation of MCMC samples
}
}
