% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcoutputpermfixpost.R
\docType{class}
\name{mcmcoutputpermfixpost-class}
\alias{mcmcoutputpermfixpost-class}
\alias{.mcmcoutputpermfixpost}
\title{Finmix \code{mcmcoutput} class for fixed indicators and posterior parameters}
\description{
This class defines the storage of parameters of the posterior distribution.
It inherits from the
}
\seealso{

}
