% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{hasWeight,model-method}
\alias{hasWeight,model-method}
\alias{simulate,model-method}
\alias{model_class}
\title{Getter for weights}
\usage{
\S4method{hasWeight}{model}(object, verbose = FALSE)

\S4method{simulate}{model}(model, N = 100, varargin, seed = 0)

\S4method{getDist}{model}(object)

\S4method{getR}{model}(object)

\S4method{getK}{model}(object)

\S4method{getWeight}{model}(object)

\S4method{getPar}{model}(object)

\S4method{getIndicmod}{model}(object)

\S4method{getIndicfix}{model}(object)

\S4method{getT}{model}(object)

\S4method{setDist}{model}(object) <- value

\S4method{setR}{model}(object) <- value

\S4method{setK}{model}(object) <- value

\S4method{setWeight}{model}(object) <- value

\S4method{setPar}{model}(object) <- value

\S4method{setIndicmod}{model}(object) <- value

\S4method{setIndicfix}{model}(object) <- value

\S4method{setT}{model}(object) <- value
}
\arguments{
\item{verbose}{A logical indicating, if the function should give a print out.}

\item{model}{An S4 model object with specified parameters and weights.}

\item{N}{An integer specifying the number of values to be simulated.}

\item{varargin}{An S4 fdata object with specified variable dimensions.}

\item{seed}{An integer specifying the seed for the RNG.
\code{r} and repetitions \code{T}.}
}
\value{
Matrix of weights.

An S4 fdata object holding the simulated values.
}
\description{
\code{hasWeight} returns the weight matrix.

\code{simulate} simulates values for a specified mixture model in an
S4 \code{model} object.
}
\section{Functions}{
\itemize{
\item \code{simulate,model-method}: Simulates data from a finite mixture model
}}

\examples{
\dontrun{
weight <- hasWeight(model)
}
\dontrun{
f_data <- simulate(model, 100)
}
}
\seealso{
\code{model}, \code{fdata}
}
