% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normultmodelmoments.R
\docType{class}
\name{normultmodelmoments}
\alias{normultmodelmoments}
\alias{.normultmodelmoments}
\title{Finmix \code{normultmodelmoments} class}
\description{
Defines a class that holds modelmoments for a finite mixture of normult
distributions. Note that this class is not directly used, but indirectly
when calling the \code{modelmoments} constructor \code{\link[=modelmoments]{modelmoments()}}.
}
\section{Slots}{

\describe{
\item{\code{B}}{A numeric defining the between-group heterogeneity.}

\item{\code{W}}{A numeric defining the within-group heterogeneity.}

\item{\code{R}}{A numeric defining the coefficient of determination.}
}}

\seealso{
\itemize{
\item \link{modelmoments_class} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
}
}
