% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\docType{class}
\name{prior-class}
\alias{prior-class}
\alias{.prior}
\alias{hasPriorPar,prior,model-method}
\alias{hasPriorWeight,prior,model-method}
\title{Finmix \code{prior} class}
\usage{
\S4method{hasPriorPar}{prior,model}(object, model, verbose = FALSE)

\S4method{hasPriorWeight}{prior,model}(object, model, verbose = FALSE)
}
\arguments{
\item{object}{A \code{prior} object containing the specifications for the prior.}

\item{model}{A \code{model} object containing the specifications for the model.}

\item{verbose}{A logical indicating, if the output should be verbose.}
}
\description{
The \code{prior} class stores the specifications for the prior distribution used
for Bayesian estimation of the finite mixture parameters and weights. There
exists next to the general constructor also an advanced constructor that
specifies a data dependent prior. See \code{\link[=priordefine]{priordefine()}} for this advanced
constructor.

Calling \code{\link[=hasPriorPar]{hasPriorPar()}} checks if \code{model}-appropriate parameters are stored
in the \code{prior} object.

Calling \code{\link[=hasPriorWeight]{hasPriorWeight()}} checks if \code{model}-appropriate weight parameters
are stored in the \code{prior} object.
}
\section{Functions}{
\itemize{
\item \code{hasPriorPar,prior,model-method}: Checks for parameters in \code{prior} object

\item \code{hasPriorWeight,prior,model-method}: Checks for prior weights in \code{prior} object
}}

\section{Slots}{

\describe{
\item{\code{weight}}{A matrix storing the prior parameters for the \code{weight} of a
finite mixture model.}

\item{\code{par}}{A list storing the prior parameters for the parameters of a finite
mixture model.}

\item{\code{type}}{A character specifying what type of prior should be used in
Bayesian estimation. Either \code{"independent"} for an independent prior
distribution or \code{"condconjugate"} for a conditionally conjugate prior
distribution.}

\item{\code{hier}}{A logical defining, if the used prior should be hierarchical.
Hierarchical prior are often more robust, but need an additional layer in
sampling, so computing costs increase.}
}}

\examples{
# Define a Poisson mixture model.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Call the default constructor.
f_prior <- prior()
# Check if the prior has appropriate parameters defined.
hasPriorPar(f_prior, f_model)
hasPriorPar(f_prior, f_model, TRUE)

# Define a Poisson mixture model.
f_model <- model("poisson", par = list(lambda = c(0.3, 0.7)), K = 2)
# Call the default constructor.
f_prior <- prior()
# Check if the prior has appropriate parameters defined.
hasPriorWeight(f_prior)
hasPriorWeight(f_prior, TRUE)

}
\references{
\itemize{
\item Frühwirth-Schnatter, S (2006), "Finite Mixture and Markov Switching Models"
}
}
\seealso{
\itemize{
\item \code{\link{prior}} for the general constructor of this class
\item \code{\link{priordefine}} for the advanced constructor of this class
}

\itemize{
\item \link[=prior-class]{prior} for the definition of the \code{prior} class
\item \link[=model_class]{model} for the definition of the \code{model} class
}

\itemize{
\item \link[=prior-class]{prior} for the definition of the \code{prior} class
\item \link[=model_class]{model} for the definition of the \code{model} class
}
}
