% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdatamoments.R
\name{show,sdatamoments-method}
\alias{show,sdatamoments-method}
\title{Shows a summary of an \code{sdatamoments} object.}
\usage{
\S4method{show}{sdatamoments}(object)
}
\arguments{
\item{object}{An \code{sdatamoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{sdatamoments} object gives an overview
of the moments of a finite mixture with discrete data.
}
\section{Functions}{
\itemize{
\item \code{show,sdatamoments-method}: Shows a summary of an object
}}

