% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalmodelmoments.R
\name{show,normalmodelmoments-method}
\alias{show,normalmodelmoments-method}
\title{Shows a summary of an \code{normalmodelmoments} object.}
\usage{
\S4method{show}{normalmodelmoments}(object)
}
\arguments{
\item{object}{An \code{normalmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{normalmodelmoments} object gives an overview
of the moments of an normal finite mixture.
}
