% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonmodelmoments.R
\name{show,poissonmodelmoments-method}
\alias{show,poissonmodelmoments-method}
\title{Shows a summary of an \code{poissonmodelmoments} object.}
\usage{
\S4method{show}{poissonmodelmoments}(object)
}
\arguments{
\item{object}{An \code{poissonmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{poissonmodelmoments} object gives an overview
of the moments of an poisson finite mixture.
}
\seealso{
\itemize{
\item \linkS4class{modelmoments} for the base class for model moments
\item \code{\link[=modelmoments]{modelmoments()}} for the constructor of \code{modelmoments} classes
}
}
\keyword{internal}
