% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{simulate}
\alias{simulate}
\title{Simulates data from a finite mixture model}
\usage{
simulate(model, N = 100, varargin, seed = 0)
}
\description{
Simulates data from a finite mixture model
}
\keyword{internal}
