% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studmultmodelmoments.R
\name{show,studmultmodelmoments-method}
\alias{show,studmultmodelmoments-method}
\title{Shows a summary of an \code{studmultmodelmoments} object.}
\usage{
\S4method{show}{studmultmodelmoments}(object)
}
\arguments{
\item{object}{An \code{studmultmodelmoments} object.}
}
\value{
A console output listing the slots and summary information about
each of them.
}
\description{
Calling \code{\link[=show]{show()}} on an \code{studmultmodelmoments} object gives an overview
of the moments of an studmult finite mixture.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Shows a summary of an object
}}

