% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{swapElements}
\alias{swapElements}
\title{Swaps elements in the MCMC sample arrays}
\usage{
swapElements(object, index)
}
\description{
Swaps elements in the MCMC sample arrays
}
\keyword{internal}
